\docType{methods}
\name{addDim}
\alias{addDim}
\alias{addDim,BrainSpace,numeric-method}
\title{Generic function to add a dimension to an object}
\usage{
addDim(x, n)

\S4method{addDim}{BrainSpace,numeric}(x, n)
}
\arguments{
  \item{x}{a dimensioned object}

  \item{n}{the size of the dimension to add}

  \item{x}{an \code{\linkS4class{BrainSpace}} object}

  \item{n}{the size of the dimension to be added}
}
\description{
Generic function to add a dimension to an object

function is used to add a dimension to a
\code{\linkS4class{BrainSpace}} object
}

