\name{AFNIFileDescriptor-class}
\alias{AFNIFileDescriptor-class}
\alias{AFNIFileDescriptor}
\title{This class supports the AFNI file format...}
\description{This class supports the AFNI file format}
\section{Extends}{\code{\linkS4class{BrainFileDescriptor}}}
\section{Methods}{\describe{\item{\code{\link[=readMetaInfo,AFNIFileDescriptor-method]{readMetaInfo}}}{}}}

