\name{AFNIMetaInfo}
\alias{AFNIMetaInfo}
\title{Constructor for \linkS4class{AFNIMetaInfo} class...}
\usage{AFNIMetaInfo(descriptor, afni_header)
}
\description{Constructor for \code{\linkS4class{AFNIMetaInfo}} class}
\value{an instance of class \code{\linkS4class{AFNIMetaInfo}}}
\arguments{\item{descriptor}{an instance of class \code{\linkS4class{AFNIFileDescriptor}}}
\item{afni_header}{a \code{list} returned by \code{readAFNIHeader}}
}

