\name{ANALYZEFile}
\alias{ANALYZEFile}

\title{ function to create a new "ANALYZEFile" instance }
\description{
  function to create a new "ANALYZEFile" instance
}
\usage{
ANALYZEFile(path, open = "r", type = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{path}{  the file path of the image }
  \item{open}{  open for input "r" or output "w" }
  \item{type}{  the fileType, "analyze-standard" or "analyze-gz" }
}
\value{
  an instance of class "ANALYZEFile"
}
\examples{
}

\keyword{manip}

