% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainData.R
\docType{methods}
\name{as.array,BrainData-method}
\alias{as.array,BrainData-method}
\title{convert \code{BrainData} instance to array}
\usage{
\S4method{as.array}{BrainData}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert \code{BrainData} instance to array
}

