% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{as.list,SparseBrainVector-method}
\alias{as.list,SparseBrainVector-method}
\title{as.list}
\usage{
\S4method{as.list}{SparseBrainVector}(x)
}
\arguments{
\item{x}{object to be coerced or tested.}
}
\description{
convert SparseBrainVector to list of DenseBrainVolumes
}

