% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{as.logical,BrainVolume-method}
\alias{as.logical,BrainVolume-method}
\title{as.logical}
\usage{
\S4method{as.logical}{BrainVolume}(x)
}
\arguments{
\item{x}{object to be coerced or tested.}
}
\description{
Convert BrainVolume to logical vector
}

