% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainData.R
\docType{methods}
\name{as.matrix,BrainData-method}
\alias{as.matrix,BrainData-method}
\title{convert \code{BrainData} instance to matrix}
\usage{
\S4method{as.matrix}{BrainData}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert \code{BrainData} instance to matrix
}

