\docType{methods}
\name{as.raster,Layer-method}
\alias{as.raster,Layer-method}
\title{as.raster}
\usage{
\S4method{as.raster}{Layer}(x, zpos, thresh = c(0, 0), axis = 3)
}
\arguments{
  \item{x}{the layer to convert}

  \item{zpos}{the z axis??}

  \item{thresh}{the threshold range}

  \item{axis}{the axis index (1,2,3)}
}
\description{
as.raster
}

