% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrainData.R
\docType{methods}
\name{as.vector,BrainData,ANY-method}
\alias{as.vector,BrainData,ANY-method}
\title{convert \code{BrainData} instance to vector}
\usage{
\S4method{as.vector}{BrainData,ANY}(x)
}
\description{
convert \code{BrainData} instance to vector
}

