% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainData.R
\docType{methods}
\name{as.vector,BrainData-method}
\alias{as.vector,BrainData-method}
\title{convert \code{BrainData} instance to vector}
\usage{
\S4method{as.vector}{BrainData}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert \code{BrainData} instance to vector
}
