\docType{methods}
\name{axes}
\alias{axes}
\alias{axes,BrainData-method}
\alias{axes,BrainSpace-method}
\title{Generic getter function to extract image axes}
\usage{
axes(x)

\S4method{axes}{BrainSpace}(x)

\S4method{axes}{BrainData}(x)
}
\arguments{
  \item{x}{an object with a set of axes}
}
\description{
Generic getter function to extract image axes

axes

axes
}

