% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{AxisSet-class}
\alias{AxisSet-class}
\title{Base}
\description{
Virtual base class representing an ordered set of named axes.
}
\section{Slots}{

\describe{
\item{\code{ndim}}{the number of axes (or dimensions)}
}}

