\name{AxisSet1D-class}
\alias{AxisSet1D-class}
\alias{AxisSet1D}
\title{A one-dimensional axis set...}
\description{A one-dimensional axis set}
\section{Slots}{\describe{\item{\code{i}:}{}}}
\section{Extends}{\code{\linkS4class{AxisSet}}}
\section{Methods}{\describe{\item{\code{\link[=show,AxisSet1D-method]{show}}}{}}}

