% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{AxisSet2D-class}
\alias{AxisSet2D-class}
\title{A two-dimensional axis set}
\description{
A two-dimensional axis set
}
\section{Slots}{

\describe{
\item{\code{j}}{the first axis}
}}

