\name{AxisSet3D-class}
\alias{AxisSet3D-class}
\alias{AxisSet3D}
\title{A three-dimensional axis set...}
\description{A three-dimensional axis set}
\section{Slots}{\describe{\item{\code{k}:}{}}}
\section{Extends}{\code{\linkS4class{AxisSet2D}}}
\section{Methods}{\describe{\item{\code{\link[=print,AxisSet3D-method]{print}}}{}\item{\code{\link[=show,AxisSet3D-method]{show}}}{}}}

