% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{AxisSet4D-class}
\alias{AxisSet4D-class}
\title{A four-dimensional axis set}
\description{
A four-dimensional axis set
}
\section{Slots}{

\describe{
\item{\code{l}}{the fourth axis}
}}

