% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{AxisSet5D-class}
\alias{AxisSet5D-class}
\title{A five-dimensional axis set}
\description{
A five-dimensional axis set
}
\section{Slots}{

\describe{
\item{\code{m}}{the fifth axis}
}}

