% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{BaseSource-class}
\alias{BaseSource-class}
\title{BaseSource}
\description{
This is a base class to represent a data source
}
\section{Slots}{

\describe{
\item{\code{metaInfo}}{meta information for the data source}
}}

