% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrainRegion3D.R
\name{BootstrapSearchlight}
\alias{BootstrapSearchlight}
\title{Create a searchlight iterator that samples regions from within a mask.
Searchlight centers are sampled *without* replacement, but the same voxel can belong to multiple searchlight samples.
It is in the latter sense that this is a bootstrap resampling scheme.}
\usage{
BootstrapSearchlight(mask, radius, iter = 100)
}
\arguments{
\item{mask}{an image volume containing valid central voxels for roving searchlight}

\item{radius}{in mm of spherical searchlight}

\item{iter}{the total number of searchlights to sample (default is 100)}
}
\description{
Create a searchlight iterator that samples regions from within a mask.
Searchlight centers are sampled *without* replacement, but the same voxel can belong to multiple searchlight samples.
It is in the latter sense that this is a bootstrap resampling scheme.
}

