% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainSpace.R
\docType{methods}
\name{bounds}
\alias{bounds}
\alias{bounds,BrainData-method}
\alias{bounds,BrainSpace-method}
\title{Generic function to extract the spatial bounds (origin + dim * spacing) of an image
param x the object}
\usage{
bounds(x)

\S4method{bounds}{BrainSpace}(x)

\S4method{bounds}{BrainData}(x)
}
\arguments{
\item{x}{the object with \code{bounds} property}
}
\description{
Generic function to extract the spatial bounds (origin + dim * spacing) of an image
param x the object

bounds

bounds
}

