% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{BasisBrainVector-class}
\alias{BasisBrainVector-class}
\title{BasisBrainVector}
\description{
a class that stores a represents a 4-dimensional array as a set of basis functions (dictionary) and 
corresponding set of coefficients.
}
\section{Slots}{

\describe{
\item{\code{mask}}{the mask defining the sparse domain}

\item{\code{basis}}{the matrix of bases, were each column is a basis vector.}

\item{\code{coeffs}}{the \code{sparseMatrix} of coefficients}

\item{\code{map}}{instance of class \code{\linkS4class{IndexLookupVolume}} is used to map between spatial and index/row coordinates}
}}

