% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClass.R, R/BrainVector.R
\docType{class}
\name{BrainBucketSource-class}
\alias{BrainBucketSource}
\alias{BrainBucketSource-class}
\title{BrainBucketSource}
\usage{
BrainBucketSource(fileName, pattern = NULL, indices = NULL)
}
\arguments{
\item{fileName}{the name of the bucket file}

\item{pattern}{optional regular expression used to filter the sub-volumes using associated labels}

\item{indices}{optional set of sub-volume indices to load}
}
\description{
A class that is used to produce a \code{\linkS4class{BrainBucket}} instance

Constructor function for \code{\linkS4class{BrainBucketSource}} class
}
\section{Slots}{

\describe{
\item{\code{sourceList}}{a list of sources for the bucket sub-volumes}

\item{\code{cache}}{a cache used to store data in memory}
}}

