% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{BrainFileDescriptor-class}
\alias{BrainFileDescriptor-class}
\title{BrainFileDescriptor}
\description{
This class represents a neuroimaging file format
}
\section{Slots}{

\describe{
\item{\code{fileFormat}}{the name of the file format (e.g. NIfTI)}

\item{\code{headerEncoding}}{the file encoding of the header file (e.g. 'raw' for binary, 'gzip' for gz compressed')}

\item{\code{headerExtension}}{the file extension for the header file (e.g. 'nii' for NIfTI single files)}

\item{\code{dataEncoding}}{the file encoding for the data file}

\item{\code{dataExtension}}{the file extension for the data file (e.g. 'nii' for NIfTI single files)}
}}

