\name{BrainSlice-class}
\alias{BrainSlice-class}
\alias{BrainSlice}
\title{BrainVolume...}
\description{BrainVolume}
\details{Two-dimensional brain image}
\section{Extends}{\code{\linkS4class{BrainData}}, \code{\linkS4class{array}}}
\section{Methods}{\describe{\item{\code{\link[=gridToIndex,BrainSlice,matrix-method]{gridToIndex}}}{}\item{\code{\link[=indexToGrid,BrainSlice,index-method]{indexToGrid}}}{}}}

