% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSlice.R
\name{BrainSlice}
\alias{BrainSlice}
\title{BrainSlice constructor}
\usage{
BrainSlice(data, space, indices = NULL)
}
\arguments{
\item{data}{data vector or matrix}

\item{space}{an instance of class BrainSpace}

\item{indices}{linear indices corresponding to data elements}
}
\description{
BrainSlice constructor
}

