% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{BrainSource-class}
\alias{BrainSource-class}
\title{Base class for representing a data source for images. The purpose of this class is to provide a layer in between
low level IO and image loading functionality.}
\description{
Base class for representing a data source for images. The purpose of this class is to provide a layer in between
low level IO and image loading functionality.
}
\section{Slots}{

\describe{
\item{\code{metaInfo}}{meta information for the data source}
}}

