\name{BrainSource}
\alias{BrainSource}
\title{Base class for representing a data source for images. The purpose of this class is to provide a layer in between
low level IO and image loading functionality.}
\description{
  Base class for representing a data source for images. The
  purpose of this class is to provide a layer in between
  low level IO and image loading functionality.
}

