% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{BrainSurfaceSource-class}
\alias{BrainSurfaceSource-class}
\title{BrainSurfaceSource}
\description{
A class that is used to produce a \code{\linkS4class{BrainSurface}} instance
}
\section{Slots}{

\describe{
\item{\code{sourceList}}{a list of sources for the bucket sub-volumes}

\item{\code{cache}}{a cache used to store data in memory}
}}

