% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{BrainSurfaceVector-class}
\alias{BrainSurfaceVector-class}
\title{BrainSurfaceVector}
\description{
a three-dimensional surface consisting of a set of triangle vertices with multiple values per vertex.
}
\section{Slots}{

\describe{
\item{\code{source}}{the data source for the surface}

\item{\code{geometry}}{the surface geometry, an instance of \code{SurfaceGeometry}}

\item{\code{indices}}{a vector of indices specifying the valid surface nodes.}

\item{\code{data}}{a \code{Matrix} of values where each column contains a vector of values 
over all the surface nodes.}
}}

