% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\name{BrainSurfaceVector}
\alias{BrainSurfaceVector}
\title{construct a new BrainSurfaceVector}
\usage{
BrainSurfaceVector(geometry, indices, mat)
}
\arguments{
\item{geometry}{a \code{SurfaceGeometry} instance}

\item{indices}{a vector of indices specifying the valid surface nodes.}

\item{mat}{a \code{matrix} of data values (rows=nodes, columns=variables)}
}
\description{
construct a new BrainSurfaceVector
}

