% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{BrainSurfaceVectorSource-class}
\alias{BrainSurfaceVectorSource-class}
\title{BrainSurfaceVectorSource}
\description{
A class that is used to produce a \code{\linkS4class{BrainSurfaceVectorSource}} instance
}
\section{Slots}{

\describe{
\item{\code{indices}}{the index vector of the volumes to be loaded}
}}

