% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R, R/BrainVector.R
\docType{class}
\name{BrainVector-class}
\alias{BrainVector-class}
\alias{BrainVector}
\title{BrainVector}
\usage{
BrainVector(data, space = NULL, mask = NULL, source = NULL, label = "")
}
\arguments{
\item{data}{the image data which can be a \code{matrix}, a 4d \code{array}, or a list of \code{BrainVolumes}. 
If the latter, the geometric space of the data \code{BrainSpace} will be inferred from the constituent volumes, 
which must all be identical.}

\item{space}{a \code{\linkS4class{BrainSpace}} object. Does not ned to be included if \code{data} argument is a list of \code{BrainVolumes}}

\item{mask}{an optional \code{array} of type \code{logical}}

\item{source}{an optional \code{\linkS4class{BrainSource}} object}

\item{label}{a label of type \code{character}}
}
\value{
a concrete instance of \code{\linkS4class{BrainVector}} class. 
If \code{mask} is provided then \code{\linkS4class{SparseBrainVector}}, otherwise \code{\linkS4class{DenseBrainVector}}
}
\description{
Four-dimensional brain image

constructor function for virtual class \code{\linkS4class{BrainVector}}
}
