% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseBrainVector.R, R/BrainVector.R
\docType{methods}
\name{[[,SparseBrainVector,numeric,ANY-method}
\alias{[[,SparseBrainVector,numeric,ANY-method}
\alias{[[,BrainVector,numeric,ANY-method}
\title{extractor}
\usage{
\S4method{[[}{SparseBrainVector,numeric,ANY}(x, i)

\S4method{[[}{BrainVector,numeric,ANY}(x, i)
}
\arguments{
\item{i}{the volume index}

\item{i}{the volume index}
}
\description{
extractor

extractor
}
