\name{BrainVector}
\alias{BrainVector}
\title{ a function to create instance of class "BrainVector" }
\description{
   a function to create instance of class "BrainVector" 
}
\usage{
BrainVector(data, space, indices = NULL)
}

\arguments{
  \item{data}{ a numeric vector containing image data }
  \item{space}{ object of class \code{BrainSpace} describing the
    geometry of the data object }
  \item{indices}{ optional argument specifiying 1-d indices if data
    vector is sparse }
}


\author{}
\examples{
}

\keyword{manip}

