% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{BrainVectorSource-class}
\alias{BrainVectorSource-class}
\title{BrainVectorSource}
\description{
A class that is used to produce a \code{\linkS4class{BrainVectorSource}} instance
}
\section{Slots}{

\describe{
\item{\code{indices}}{the index vector of the volumes to be loaded}
}}

