% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BrainVector.R
\name{BrainVectorSource}
\alias{BrainVectorSource}
\title{BrainVectorSource}
\usage{
BrainVectorSource(fileName, indices = NULL, mask = NULL)
}
\arguments{
\item{fileName}{name of the 4-dimensional image file}

\item{indices}{the subset of integer volume indices to load -- if \code{NULL} then all volumes will be loaded}

\item{mask}{image volume indicating the subset of voxels that will be loaded. If provided, function returns \code{\linkS4class{SparseBrainVectorSource}}}
}
\value{
a instance deriving from \code{\linkS4class{BrainVectorSource}}
}
\description{
Construct a \code{\linkS4class{BrainVectorSource}} object
}
\details{
If a \code{mask} is supplied then it should be a \code{\linkS4class{LogicalBrainVolume}} or \code{\linkS4class{BrainVolume}} instance. If the latter, then the mask will be defined by nonzero elements of the volume.
}

