% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainVolume.R
\docType{methods}
\name{clusterCenters}
\alias{clusterCenters}
\alias{clusterCenters,ClusteredBrainVolume,matrix,missing-method}
\title{clusterCenters}
\usage{
clusterCenters(x, features, FUN)

\S4method{clusterCenters}{ClusteredBrainVolume,matrix,missing}(x, features)
}
\arguments{
\item{x}{the object to extract cluster centers from}

\item{features}{additional features}

\item{FUN}{a user-supplied function}
}
\description{
clusterCenters
}
