% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainRegion3D.R
\name{ClusteredSearchlight}
\alias{ClusteredSearchlight}
\title{Create a clustered Searchlight iterator}
\usage{
ClusteredSearchlight(mask, csize)
}
\arguments{
\item{mask}{an image volume containing valid central voxels for roving searchlight}

\item{csize}{the number of clusters}
}
\value{
an \code{iter} class
}
\description{
Create a clustered Searchlight iterator
}

