% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ColumnReader-class}
\alias{ColumnReader-class}
\title{ColumnReader}
\description{
This class supports reading of data froma matrix-like stroage format
}
\section{Slots}{

\describe{
\item{\code{nrow}}{the number of rows}

\item{\code{ncol}}{the number of columns}

\item{\code{reader}}{a function that takes a set of column indices and returns a \code{matrix}}
}}

