% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryIO.R
\name{ColumnReader}
\alias{ColumnReader}
\title{ColumnReader}
\usage{
ColumnReader(nrow, ncol, reader)
}
\arguments{
\item{nrow}{the number of rows}

\item{ncol}{the number of columns}

\item{reader}{a function that takes a set of column indices and returns a \code{matrix}}
}
\description{
Constructor for  \code{\linkS4class{ColumnReader}} class
}

