\name{concat,DenseBrainVolume,DenseBrainVolume-method}
\alias{concat,DenseBrainVolume,DenseBrainVolume-method}
\alias{concat}
\title{concatenate two BrainVolumes...}
\description{concatenate two BrainVolumes}
\note{dimensions of x and y must be equal}
\arguments{\item{x}{the first volume}
\item{y}{the second volume}
\item{...}{extra arguments of class BrainVolume or other concatenable objects}
}

