% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainVector.R, R/BrainVolume.R, R/SparseBrainVector.R
\docType{methods}
\name{concat}
\alias{concat}
\alias{concat,BrainVector,BrainVector-method}
\alias{concat,BrainVector,BrainVolume-method}
\alias{concat,BrainVolume,BrainVector-method}
\alias{concat,DenseBrainVolume,DenseBrainVolume-method}
\alias{concat,SparseBrainVector,SparseBrainVector-method}
\title{Concatenate two objects}
\usage{
concat(x, y, ...)

\S4method{concat}{BrainVector,BrainVolume}(x, y, ...)

\S4method{concat}{BrainVolume,BrainVector}(x, y, ...)

\S4method{concat}{BrainVector,BrainVector}(x, y, ...)

\S4method{concat}{DenseBrainVolume,DenseBrainVolume}(x, y, ...)

\S4method{concat}{SparseBrainVector,SparseBrainVector}(x, y, ...)
}
\arguments{
\item{x}{the first object}

\item{y}{the second object}

\item{...}{additional objects}
}
\description{
Concatenate two objects

concatenate two BrainVolumes

concat
}
\note{
dimensions of x and y must be equal
}

