\docType{methods}
\name{coords}
\alias{coords}
\alias{coords,IndexLookupVolume-method}
\alias{coords,ROIVolume-method}
\alias{coords,SparseBrainVector-method}
\title{Extract coordinates}
\usage{
coords(x, ...)

\S4method{coords}{ROIVolume}(x)

\S4method{coords}{IndexLookupVolume}(x, i)

\S4method{coords}{SparseBrainVector}(x, i)
}
\arguments{
  \item{x}{the object to extract coordinates from}

  \item{x}{an ROIVolume object}
}
\description{
Extract coordinates

coords

coords

coords
}

