% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainSpace.R, R/BrainVolume.R
\docType{methods}
\name{coordToGrid}
\alias{coordToGrid}
\alias{coordToGrid,BrainSpace,matrix-method}
\alias{coordToGrid,BrainSpace,numeric-method}
\alias{coordToGrid,BrainVolume,matrix-method}
\title{Generic function to convert N-dimensional real world coordinates to grid coordinates}
\usage{
coordToGrid(x, coords)

\S4method{coordToGrid}{BrainSpace,matrix}(x, coords)

\S4method{coordToGrid}{BrainSpace,numeric}(x, coords)

\S4method{coordToGrid}{BrainVolume,matrix}(x, coords)
}
\arguments{
\item{x}{the object}

\item{coords}{a matrix of real world coordinates}
}
\value{
a matrix of grid coordinates
}
\description{
Generic function to convert N-dimensional real world coordinates to grid coordinates

coordToGrid

coordToGrid
}

