\name{createSpace-methods}
\docType{methods}
\alias{createSpace-methods}
\alias{createSpace,ANALYZEHeader-method}
\alias{createSpace}
\title{creates a brainSpace instance}
\description{
 Methods for function \code{createSpace}
}
\section{Methods}{
\describe{

\item{x = "NIFTIHeader"}{  creates a \code{BrainSpace} instance}
\item{x = "ANALYZEHeader"}{ creates a \code{BrainSpace} instance}
}}
\keyword{methods}

