\name{dataDim-methods}
\docType{methods}
\alias{dataDim-methods}
\alias{dataDim,ANALYZEHeader-method}
\alias{dataDim}
\title{ method to access the data dimension from a header }
\description{
 Methods for function \code{dataDim} 
}
\section{Methods}{
\describe{

\item{x = "NIFTIHeader"}{ returns data dimension for image described by \code{x} }
\item{x = "ANALYZEHeader"}{ returns data dimension for image described by \code{x} }

}}
\keyword{methods}

