\name{dataDim}
\alias{dataDim}

\title{ retrieves the data dimension vector from header }
\description{
  generic function to access the data dimension stored in a object
  deriving from\code{BrainHeader}}
\usage{
dataDim(x)
}

\arguments{
  \item{x}{ object deriving from class "BrainHeader" }
}
\details{
  
}

\value{
  returns a vector of integers representing data dimensions.
}

\examples{



}

\keyword{ manip }

