\name{dataFile}
\alias{dataFile}

\title{ access the name of the data file associated with a "BrainHeader" object
}

\description{
  acesses the name of the file stored in a "BrainHeader" instance.
}
\usage{
dataFile(x)
}

\arguments{
  \item{x}{ object deriving from class "BrainHeader" }
}

\value{
  a character string returning the filename of the image data.
}

\examples{

}

\keyword{ manip }

