\name{dataFileExt-methods}
\docType{methods}
\alias{dataFileExt-methods}

\title{ generic method returning name of data file extension }
\description{
  returns the extension of the file containing the image data.
}
\section{Methods}{
\describe{
\item{x = "BrainFile"}{returns the extension of the supplied BrainFile object}

}}

\keyword{methods}

