\name{dataFileExt}
\alias{dataFileExt}
\title{ gets the file extension for data file }
\description{
  accessor for the file extension for data file 
}
\usage{
dataFileExt(x)
}

\arguments{
  \item{x}{ object deriving from class "BrainHeader" }
}

\value{
  character string of file extension
}

\examples{

}
\keyword{ manip }

