% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{dataFileMatches}
\alias{dataFileMatches}
\alias{dataFileMatches,BrainFileDescriptor,character-method}
\title{Generic function to test whether a file name conforms to the given a \code{\linkS4class{BrainFileDescriptor}} instance.
Will test for match to data file only}
\usage{
dataFileMatches(x, fileName)

\S4method{dataFileMatches}{BrainFileDescriptor,character}(x, fileName)
}
\arguments{
\item{x}{object for which the file name is to matched to}

\item{fileName}{file name to be matched}
}
\value{
TRUE for match, FALSE otherwise
}
\description{
Generic function to test whether a file name conforms to the given a \code{\linkS4class{BrainFileDescriptor}} instance.
Will test for match to data file only
}

