\name{dataOffset-methods}
\docType{methods}
\alias{dataOffset-methods}
\alias{dataOffset,BrainHeader-method}
\title{ Methods for the dataOffset generic }
\description{
  the offset in bytes for the image values contained in a data file
 }

\section{Methods}{
  \describe{
    \item{x = "BrainHeader"}{Returns an integer value representing the byte offset for the image file}
    }
  }

  \keyword{methods}
