\name{dataOffset}
\alias{dataOffset}
\title{ gets the data offset in bytes for an image file }
\description{
  function for accessing the offset in bytes marking the place at which
  the image data is stored in the file.
}
\usage{
dataOffset(x)
}

\arguments{
  \item{x}{ obect derving from class "BrainHeader" }
}

\value{
  an integer value for the byte offset
}


\examples{

}

\keyword{manip }
