\name{dataType}
\alias{dataType}
\title{ a string representing the data type of the image data stored on
  file }
\description{
  accessor for the string representation for the image data as it is
  stored on disk (which may be distinct from the sdtorage type in R)
}
\usage{
dataType(x)
}

\arguments{
  \item{x}{ object deriving from class "BrainHeader" }
}

\value{
  character string representing the data type
}

\examples{
}

\keyword{ manip }

