% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R, R/BrainVector.R
\docType{class}
\name{DenseBrainVector-class}
\alias{DenseBrainVector-class}
\alias{DenseBrainVector}
\title{DenseBrainVector}
\usage{
DenseBrainVector(data, space, source = NULL, label = "")
}
\arguments{
\item{data}{a 4-dimensional \code{array} or a 2-dimension \code{matrix} that is either nvoxels by ntime-points or ntime-points by nvoxels}

\item{space}{a \code{\linkS4class{BrainSpace}} object}

\item{source}{an optional \code{\linkS4class{BrainSource}} object}

\item{label}{a label of type \code{character}}
}
\value{
\code{\linkS4class{DenseBrainVector}} instance
}
\description{
Four-dimensional brain image, backed by an array

constructor function for class \code{\linkS4class{DenseBrainVector}}
}
