\name{DenseBrainVolume}
\alias{DenseBrainVolume}
\title{Construct a \linkS4class{DenseBrainVolume} instance...}
\usage{DenseBrainVolume(data, space, source)
}
\description{Construct a \code{\linkS4class{DenseBrainVolume}} instance}
\arguments{\item{data}{a three-dimensional \code{array}}
\item{space}{an instance of class \code{\linkS4class{BrainSpace}}}
\item{source}{an instance of class \code{\linkS4class{BrainSource}}}
}

