% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainData.R
\docType{methods}
\name{dim,BrainData-method}
\alias{dim,BrainData-method}
\title{dim of \code{BrainData} object}
\usage{
\S4method{dim}{BrainData}(x)
}
\arguments{
\item{x}{the object}
}
\description{
dim of \code{BrainData} object
}

