% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrainMetaInfo.R
\docType{methods}
\name{dim,FileMetaInfo-method}
\alias{dim,FileMetaInfo-method}
\title{dim of \code{FileMetaInfo}}
\usage{
\S4method{dim}{FileMetaInfo}(x)
}
\arguments{
\item{x}{the object}
}
\description{
dim of \code{FileMetaInfo}
}

