% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BrainData.R, R/BrainMetaInfo.R, R/BrainSpace.R
\docType{methods}
\name{dim,BrainData-method}
\alias{dim,BrainData-method}
\alias{dim,BrainSpace,ANY-method}
\alias{dim,BrainSpace-method}
\alias{dim,FileMetaInfo,ANY-method}
\alias{dim,FileMetaInfo-method}
\title{dim}
\usage{
\S4method{dim}{BrainData}(x)

\S4method{dim}{FileMetaInfo}(x)

\S4method{dim}{BrainSpace}(x)
}
\description{
dim

dim
}

