% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainVector.R, R/SparseBrainVector.R
\docType{methods}
\name{eachVolume}
\alias{eachVolume}
\alias{eachVolume,BrainBucket,function,logical,ANY-method}
\alias{eachVolume,BrainBucket,function,missing,missing-method}
\alias{eachVolume,BrainVector,function,logical,ANY-method}
\alias{eachVolume,BrainVector,function,missing,BrainVolume-method}
\alias{eachVolume,BrainVector,function,missing,missing-method}
\alias{eachVolume,SparseBrainVector,function,logical,missing-method}
\alias{eachVolume,SparseBrainVector,function,missing,LogicalBrainVolume-method}
\alias{eachVolume,SparseBrainVector,function,missing,missing-method}
\title{Generic function to apply a function to each volume of a four-dimensional image}
\usage{
eachVolume(x, FUN, withIndex, mask, ...)

\S4method{eachVolume}{BrainVector,`function`,missing,missing}(x, FUN, withIndex,
  mask, ...)

\S4method{eachVolume}{BrainVector,`function`,missing,BrainVolume}(x, FUN,
  withIndex, mask, ...)

\S4method{eachVolume}{BrainVector,`function`,missing,missing}(x, FUN, withIndex,
  mask, ...)

\S4method{eachVolume}{BrainBucket,`function`,missing,missing}(x, FUN, withIndex,
  mask, ...)

\S4method{eachVolume}{BrainBucket,`function`,logical,ANY}(x, FUN, withIndex,
  mask, ...)

\S4method{eachVolume}{BrainVector,`function`,logical,ANY}(x, FUN, withIndex,
  mask, ...)

\S4method{eachVolume}{SparseBrainVector,`function`,logical,missing}(x, FUN,
  withIndex = FALSE, mask, ...)

\S4method{eachVolume}{SparseBrainVector,`function`,missing,missing}(x, FUN,
  withIndex, mask, ...)


  \S4method{eachVolume}{SparseBrainVector,`function`,missing,LogicalBrainVolume}(x,
  FUN, withIndex, mask, ...)
}
\arguments{
\item{x}{four-dimensional image}

\item{FUN}{a \code{function} taking one or two arguments (depending on the value of \code{withIndex}}

\item{withIndex}{whether the index of the volume supplied as the second argument to the function}

\item{mask}{an image mask indicating subset of volume elements to apply function over}

\item{...}{additional arguments}
}
\description{
Generic function to apply a function to each volume of a four-dimensional image

eachVolume

eachVolume

eachVolume

eachVolume

eachVolume

eachVolume
}

