\name{encoding}
\alias{encoding}
\title{ retrieves the encoding method used for image data }
\description{
  accessor for the way in which the image data is encoded on disk
  (e.g. "raw", "gzip", etc.")
}
\usage{
encoding(x)
}

\arguments{
  \item{x}{ an object deriving from class "BrainHeader" }
}

\value{
  a character string representing the encoding type
}

\examples{
}

\keyword{ manip }

