\name{endian-methods}
\docType{methods}
\alias{endian-methods}
\alias{endian,BrainHeader-method}
\title{ get eendianess of data storage }
\description{
  get eendianess of data storage
}
\section{Methods}{
\describe{

\item{x = "BrainHeader"}{ object deriving from "BrainHeader" class  }
}}
\keyword{methods}

