\name{endian}
\alias{endian}

\title{ the endianess of the stored image data }
\description{
  retrieves the endianess ("big" or "little") of the binary image data
}
\usage{
endian(x)
}

\arguments{
  \item{x}{ an instance deriving from class "BrainHeader" }
}

\value{
 character string which can be either "biug" or "little"
}

\examples{
}
\keyword{ manip }

