\name{extraInfo}
\alias{extraInfo}
\title{ returns auxilliary header information }

\description{
  function to access format-specific header information that is stored
  in \code{list} object
}
\usage{
extraInfo(x)
}

\arguments{
  \item{x}{ an object derived from class "BrainHeader" }
}

\value{
a list containing all auxilliary header information
}

\examples{
}
\keyword{ manip }

