\name{fileName}
\alias{fileName}
\title{ gets the file name assocaited with a "BrainFile" object }

\description{
  gets the file name assocaited with a "BrainFile" object
}
\usage{
fileName(x)
}


\arguments{
  \item{x}{ an object deriving from then "BrainFile" class }
}

\value{
a character string for the file name
}
\examples{
}
\keyword{ manip }

