\name{fileType}
\alias{fileType}

\title{ the image file type }
\description{
  the file format used to store the image data, e.g. Analyze7.5, NIFTI,
  or AFNI
}
\usage{
fileType(x)
}

\arguments{
  \item{x}{ an object deriving from class "BrainHeader" }
}

\value{
character string of file format name
}

\examples{

}
\keyword{ manip }

