% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{FreesurferAsciiSurfaceFileDescriptor-class}
\alias{FreesurferAsciiSurfaceFileDescriptor-class}
\title{FresurferAsciiSurfaceFileDescriptor}
\description{
This class supports the FreesurferAsciiSurfaceFileDescriptor file format for surface geometry
}

