% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{FreesurferSurfaceGeometryMetaInfo-class}
\alias{FreesurferSurfaceGeometryMetaInfo-class}
\title{FreeSurferSurfaceGeometryMetaInfo
	
This class contains meta information for brain surface geometry}
\description{
FreeSurferSurfaceGeometryMetaInfo
	
This class contains meta information for brain surface geometry
}

