% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainMetaInfo.R
\name{FreesurferSurfaceGeometryMetaInfo}
\alias{FreesurferSurfaceGeometryMetaInfo}
\title{Constructor for \code{\linkS4class{SurfaceGeometryMetaInfo}} class}
\usage{
FreesurferSurfaceGeometryMetaInfo(descriptor, header)
}
\arguments{
\item{descriptor}{the file descriptor}

\item{header}{a \code{list} containing header information}
}
\description{
Constructor for \code{\linkS4class{SurfaceGeometryMetaInfo}} class
}

