% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainSurface.R
\docType{methods}
\name{graph}
\alias{graph}
\alias{graph,BrainSurface-method}
\alias{graph,BrainSurfaceVector-method}
\alias{graph,SurfaceGeometry-method}
\title{extract \code{igraph} object}
\usage{
graph(x, ...)

\S4method{graph}{BrainSurface}(x, ...)

\S4method{graph}{BrainSurfaceVector}(x, ...)

\S4method{graph}{SurfaceGeometry}(x)
}
\arguments{
\item{x}{the object to extract the graph from}

\item{...}{extra args}
}
\description{
extract \code{igraph} object
}

