% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainSpace.R
\docType{methods}
\name{gridToCoord}
\alias{gridToCoord}
\alias{gridToCoord,BrainSpace,matrix-method}
\alias{gridToCoord,BrainVolume,matrix-method}
\title{Generic function to convert N-dimensional grid coordinates to real world coordinates}
\usage{
gridToCoord(x, coords)

\S4method{gridToCoord}{BrainSpace,matrix}(x, coords)

\S4method{gridToCoord}{BrainVolume,matrix}(x, coords)
}
\arguments{
\item{x}{the object}

\item{coords}{a matrix of grid coordinates}
}
\value{
a matrix of real coordinates
}
\description{
Generic function to convert N-dimensional grid coordinates to real world coordinates
}

