% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainSpace.R
\docType{methods}
\name{gridToGrid}
\alias{gridToGrid}
\alias{gridToGrid,BrainSpace,matrix-method}
\alias{gridToGrid,matrix,matrix-method}
\title{Generic function to convert voxel coordinates in the reference space (LPI) to native array space.}
\usage{
gridToGrid(x, vox)

\S4method{gridToGrid}{BrainSpace,matrix}(x, vox)

\S4method{gridToGrid}{matrix,matrix}(x, vox)
}
\arguments{
\item{x}{the object}

\item{coords}{a matrix of LPI voxel coordinates}
}
\value{
a matrix of native voxel coordinates
}
\description{
Generic function to convert voxel coordinates in the reference space (LPI) to native array space.
}
