% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainFileDescriptor.R
\docType{methods}
\name{headerFile}
\alias{headerFile}
\alias{headerFile,BrainFileDescriptor,character-method}
\title{Generic function to get the name of the header file, given a file name and a \code{\linkS4class{BrainFileDescriptor}} instance.}
\usage{
headerFile(x, fileName)

\S4method{headerFile}{BrainFileDescriptor,character}(x, fileName)
}
\arguments{
\item{x}{descriptor instance}

\item{fileName}{file name to be stripped of its extension}
}
\value{
the correct header name
}
\description{
Generic function to get the name of the header file, given a file name and a \code{\linkS4class{BrainFileDescriptor}} instance.
}
