\name{headerFile}
\alias{headerFile}

\title{ gets the name of the header file associated with a "BrainHeader" instance }
\description{
  simple accessor fucntion that retrieves the name of the header file
  assocaited a "BrainHeader" instance 
}
\usage{
headerFile(x)
}

\arguments{
  \item{x}{ object deriving from class "BrainHeader" }
}

\value{
  character string of the the full path of the header name 
}

\examples{
}
\keyword{ manip }

