\name{headerFileExt-methods}
\docType{methods}
\alias{headerFileExt-methods}
\title{ retrieves the header file extension }
\description{
 function that returns the header extension for a FIFTIFile obejct
}
\section{Methods}{
\describe{

\item{x = "NIFTIFile"}{ an object of the "NIFTIFile" class }
}}
\keyword{methods}

