\name{headerFileExt}
\alias{headerFileExt}
\title{ gets the header file extension }
\description{
  generic function to access extension (e.g. "hdr", ".nii", ".HEAD")
  appropriate for file type
  
}
\usage{
headerFileExt(x)
}

\arguments{
  \item{x}{ object derived from class "BrainHeader" }
}

\value{
character string representing file extension
}


\examples{
}
\keyword{ manip }

