% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Display.R
\name{imageGrid}
\alias{imageGrid}
\title{imageGrid}
\usage{
imageGrid(layer, gridDim = c(3, 3), zstart, zend, panelSize = 3,
  panelUnit = "inches", interpolate = FALSE, fontCol = "red")
}
\description{
imageGrid
}

