\name{index-class}
\docType{class}
\alias{index-class}

\title{Virtual Class "index" - Simple Class for image indices }
\description{ The class \code{"index"} is a virtual class used for
  indices (in signatures) for image data indexing and sub-assignment of
  brain image data.   }
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\examples{
showClass("index")
}
\keyword{classes}
