% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClass.R, R/IndexLookupVolume.R
\docType{class}
\name{IndexLookupVolume-class}
\alias{IndexLookupVolume}
\alias{IndexLookupVolume-class}
\title{IndexLookupVolume}
\usage{
IndexLookupVolume(space, indices)
}
\arguments{
\item{space}{a BrainSpace object}

\item{indices}{the set of 1-d indices defining the lookup map}
}
\description{
Three-dimensional brain image that can be used as a map between 1D grid indices and a table of values
Currently used in the \code{\linkS4class{SparseBrainVector}} class.

IndexLookupVolume
}

