\name{IndexLookupVolume}
\Rdversion{1.1}
\alias{IndexLookupVolume}
\docType{data}
\title{
 creation function for IndexLookupVolume class 
}
\description{
 this function creates an instance of class \code{IndexLookupVolume}
}

\usage{
IndexLookupVolume(space, indices)
}

\arguments{
	\item{space}{ the \code{BrainSpace} of the volume to be created }
	\item{indices}{ the 1-d indices that map to the lookup table }
}


