\name{indexToVoxel}
\alias{indexToVoxel}
\title{ converts between a 1D index and a multidimensional grid location
}
\description{
  function to converts between a 1D index and a multidimensional grid location
}
\usage{
indexToVoxel(x, idx)
}


\arguments{
  \item{x}{ an object derived from class "BrainData" or "BrainSpace" }
  \item{idx}{ the 1D interger-valued index }
}

\value{
  the grid location corresponding to the index
}

\examples{

}

\keyword{ manip }

