% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{.isExtension}
\alias{.isExtension}
\alias{.concat4D}
\alias{.extract.array}
\alias{.gridToIndex3D}
\alias{.gridToIndex}
\alias{.indexToGrid}
\alias{.getRStorage}
\alias{.getDataStorage}
\alias{.getDataCode}
\alias{.getDataSize}
\alias{.getEndian}
\alias{.niftiExt}
\alias{.matrixToQuatern}
\title{.isExtension}
\usage{
.isExtension(fname, extension)

.concat4D(x, y, ...)

.extract.array(x, ..., drop = FALSE, indices = list(...))

.gridToIndex3D(dimensions, voxmat)

.gridToIndex(dimensions, vmat)

.indexToGrid(idx, array.dim)

.getRStorage(dataType)

.getDataStorage(code)

.getDataCode(dataType)

.getDataSize(dataType)

.getEndian(conn)

.niftiExt(filetype)

.matrixToQuatern(mat)
}
\description{
.isExtension

.concat4D

.extract.array

.gridToIndex3D

.gridToIndex

.indexToGrid

.getRStorage

.getDataStorage

.getDataCode

.getDataSize

.getEndian
}
\keyword{internal}
