% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{inverseTrans}
\alias{inverseTrans}
\alias{inverseTrans,BrainData-method}
\alias{inverseTrans,BrainSpace-method}
\title{Generic getter to extract inverse image coordinate transformation}
\usage{
inverseTrans(x)

\S4method{inverseTrans}{BrainSpace}(x)

\S4method{inverseTrans}{BrainData}(x)
}
\arguments{
\item{x}{an object}
}
\description{
Generic getter to extract inverse image coordinate transformation

inverseTrans

inverseTrans
}

