\name{invTrans}
\alias{invTrans}
\title{ inverse cooordinate transform }
\description{
  a matrix representing the inverse transform from real world coordinate
  to image grid space
}
\usage{
invTrans(x)
}

\arguments{
  \item{x}{ a object derived from class "BrainSpace" }
}

\value{

}

\examples{
}

\keyword{ manip }

