% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{Layer-class}
\alias{Layer-class}
\title{Layer}
\description{
A class used for displaying 2D images with color maps
}
\section{Slots}{

\describe{
\item{\code{vol}}{the \code{\linkS4class{BrainVolume}} that provides the data for the layer.}

\item{\code{colorMap}}{a character vector of colors in hexadecimal rgb format. 
Can be generated by calls to \code{rainbow}, \code{heat.colors}, \code{topo.colors}, \code{terrain.colors} or similar functions.}

\item{\code{thresh}}{cut-off value above which pixels will be made transparent.}

\item{\code{view}}{the anatomical orientation of the layer view: type \code{AxisSet3D}}

\item{\code{zero_col}}{the color of pixels with intensity of zero. This value overrides the color from the slot \code{colorMap}}

\item{\code{alpha}}{the transparency of the layer (0 to 1)}
}}

