% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\docType{data}
\name{Layer}
\alias{Layer}
\title{Layer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Layer
}
\arguments{
\item{vol}{volume instance of \code{\linkS4class{BrainVolume}}}

\item{colorMap}{a lookup table defining mapping from image intensity values to colors.}

\item{thresh}{a range (min,max) defining the threshold window for determining image opacity.}

\item{axis}{the axis index of the axis perpendicular to the xy plane (options: 1,2,3; default is 3)}

\item{zero_col}{the color used when the value is zero.}

\item{alpha}{transparency multiplier, vlaue between 0 and 1.}
}
\value{
an object of class \code{Layer}
}
\description{
create a \code{\linkS4class{Layer}} object
}
\keyword{datasets}
