% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BrainVector.R
\docType{methods}
\name{length,BrainVector-method}
\alias{length,BrainVector-method}
\title{Get length of \code{BrainVector}. This is the numbe rof volumes in the volume vector (e.g. the 4th image dimension)}
\usage{
\S4method{length}{BrainVector}(x)
}
\description{
Get length of \code{BrainVector}. This is the numbe rof volumes in the volume vector (e.g. the 4th image dimension)
}

