% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BrainVector.R
\name{loadBucket}
\alias{loadBucket}
\title{loadBucket}
\usage{
loadBucket(fileName, pattern = NULL, indices = NULL)
}
\arguments{
\item{fileName}{the name of the file to load}

\item{pattern}{optional regular expression used to filter the sub-volumes using associated labels}

\item{indices}{optional set of sub-volume indices to load}
}
\description{
load a BrainBucket object from file
}

