\name{loadData,BrainBucketSource-method}
\alias{loadData,BrainBucketSource-method}
\alias{loadData}
\title{Load data from a \linkS4class{BrainBucketSource}...}
\description{Load data from a \code{\linkS4class{BrainBucketSource}}}
\value{an instance of class \code{\linkS4class{BrainVolume}}}
\arguments{\item{x}{an instance of class \code{\linkS4class{BrainBucketSource}}}
}

