\name{loadData,BrainVectorSource-method}
\alias{loadData,BrainVectorSource-method}
\alias{loadData}
\title{Load data from a \linkS4class{BrainVectorSource}...}
\description{Load data from a \code{\linkS4class{BrainVectorSource}}}
\value{an instance of class \code{\linkS4class{BrainVector}}}
\arguments{\item{x}{an instance of class \code{\linkS4class{BrainVectorSource}}}
}

