% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainBucket.R,
%   R/SparseBrainVector.R, R/BrainVector.R, R/BrainVolume.R
\docType{methods}
\name{loadData}
\alias{loadData}
\alias{loadData,BrainBucketSource-method}
\alias{loadData,SparseBrainVectorSource-method}
\alias{loadData,BrainVectorSource-method}
\alias{loadData,BrainVolumeSource-method}
\title{Generic function to load data from a data source}
\usage{
loadData(x, ...)

\S4method{loadData}{BrainBucketSource}(x, key)

\S4method{loadData}{SparseBrainVectorSource}(x, mmap = FALSE)

\S4method{loadData}{BrainVectorSource}(x, mmap = FALSE)

\S4method{loadData}{BrainVolumeSource}(x)
}
\arguments{
\item{x}{a data source}

\item{...}{additional arguments}

\item{key}{the name or index of the bucket to load}

\item{mmap}{use memory-mapped file}
}
\value{
an instance of class \code{\linkS4class{BrainVolume}}

an instance of class \code{\linkS4class{BrainVector}}
}
\description{
Generic function to load data from a data source

Load data from a \code{\linkS4class{BrainBucketSource}}

loadData

load a BrainVolume
}
