% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\name{loadFSSurface}
\alias{loadFSSurface}
\title{load Freesurfer ascii surface}
\usage{
loadFSSurface(mesh)
}
\arguments{
\item{mesh}{file name of mesh to read in.}
}
\description{
load Freesurfer ascii surface
}

