% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainVolume.R
\name{loadVolume}
\alias{loadVolume}
\title{Load an image volume from a file}
\usage{
loadVolume(fileName, index = 1)
}
\arguments{
\item{fileName}{the name of the file to load}

\item{index}{the index of the volume (e.g. if the file is 4-dimensional)}
}
\value{
an instance of the class \code{\linkS4class{DenseBrainVolume}}
}
\description{
Load an image volume from a file
}
\examples{
fname <- system.file("extdata", "global_mask.nii", package="neuroim")
x <- loadVolume(fname)
print(dim(x))
space(x)

}
