% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/SparseBrainVector.R, R/IndexLookupVolume.R
\docType{methods}
\name{lookup}
\alias{lookup}
\alias{lookup,IndexLookupVolume,numeric-method}
\alias{lookup,SparseBrainVector,numeric-method}
\title{Index Lookup operation}
\usage{
lookup(x, i, ...)

\S4method{lookup}{SparseBrainVector,numeric}(x, i)

\S4method{lookup}{IndexLookupVolume,numeric}(x, i)
}
\arguments{
\item{x}{the object to query}

\item{i}{the index to lookup}

\item{...}{additional arguments}
}
\description{
Index Lookup operation

lookup
}

