% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{map}
\alias{map}
\alias{map,BrainVolume,Kernel-method}
\title{Generic function to apply a function to an object}
\usage{
map(x, m, ...)

\S4method{map}{BrainVolume,Kernel}(x, m, mask = NULL)
}
\arguments{
\item{x}{the object that is mapped}

\item{m}{the mapping object}

\item{...}{additional arguments}

\item{mask}{restrict application of kernel to maksed area}
}
\description{
Generic function to apply a function to an object

apply a kernel function to a BrainVolume
}

