% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{mapToColors}
\alias{mapToColors}
\title{mapToColors}
\usage{
mapToColors(imslice, col = heat.colors(128, alpha = 1),
  zero.col = "#00000000")
}
\arguments{
\item{imslice}{an image matrix defining intensity values}

\item{col}{a color map}

\item{zero.col}{the background color.}
}
\description{
map an matrix of intensity values to a matrix of color values.
}

