% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Axis.R
\name{matchAnatomy2D}
\alias{matchAnatomy2D}
\title{given two named axes return AxisSet2D singleton}
\usage{
matchAnatomy2D(axis1, axis2)
}
\arguments{
\item{axis1}{the first axis}

\item{axis2}{the second axis}
}
\description{
given two named axes return AxisSet2D singleton
}
