\docType{methods}
\name{mergePartitions,ClusteredBrainVolume,numeric,matrix-method}
\alias{mergePartitions,ClusteredBrainVolume,numeric,matrix-method}
\title{merge partititons in a ClusteredBrainVolume}
\usage{
\S4method{mergePartitions}{ClusteredBrainVolume,numeric,matrix}(x, K, features)
}
\description{
merge partititons in a ClusteredBrainVolume
}

