\name{mergePartitions}
\alias{mergePartitions}
\title{mergePartitions}
\usage{
mergePartitions(x, K, features, ...)
}
\arguments{
  \item{x}{the object to merge}

  \item{K}{the number of merged partitions}

  \item{features}{the features used to define the
  partition}
}
\description{
mergePartitions
}

