% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\name{meshToGraph}
\alias{meshToGraph}
\title{construct a graph from a set of vertices and nodes faces.}
\usage{
meshToGraph(vertices, nodes)
}
\arguments{
\item{N-by-3}{matrix of vertices}

\item{matrix}{of node faces, where each row is a set of three vertex indices.}
}
\value{
an \code{igraph} instance representing th mesh connectivity.
}
\description{
construct a graph from a set of vertices and nodes faces.
}

