% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{MNI_SPACE_1MM}
\alias{MNI_SPACE_1MM}
\title{MNI SPACE 1MM}
\format{a \code{BrainSpace} instance}
\usage{
data(MNI_SPACE_1MM)
}
\description{
This is a \code{BrainSpace} object encoding the geometry of the MNI_1MM template.
}
\keyword{datasets}

