% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{NamedAxis-class}
\alias{NamedAxis-class}
\title{NamedAxis}
\description{
This class represents an axis with a name attribute
}
\section{Slots}{

\describe{
\item{\code{axis}}{the name of the axis}

\item{\code{direction}}{of axis (-1,+1)}
}}

