% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrainVector.R
\docType{methods}
\name{names,BrainBucketSource-method}
\alias{names,BrainBucket-method}
\alias{names,BrainBucketSource-method}
\title{names}
\usage{
\S4method{names}{BrainBucketSource}(x)

\S4method{names}{BrainBucket}(x)
}
\arguments{
\item{x}{the object to get \code{names} of}
}
\description{
names
}

