\name{neuroim-package}
\alias{neuroim-package}
\alias{neuroim-package}
\title{a basic package for working with neuroimaging data}
\description{A basic package for working with neuroimaging data}
\details{\tabular{ll}{
Package: \tab neuroim\cr
Version: \tab 0.5.8\cr
Date: \tab 2010-07-10\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr}
\alias{neuroim}
\docType{package}
\author{Bradley Buchsbaum \email{bbuchsbaum@rotman-baycrest.on.ca}}
\keyword{packages}

