\name{NIFTIFile}
\alias{NIFTIFile}
\title{ function to create a new "NIFTIFile" instance }
\description{
  function to create a new "NIFTIFile" instance
}
\usage{
NIFTIFile(path, open = "r", type = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{path}{ the file path of the image }
  \item{open}{ open for input "r" or output "w" }
  \item{type}{ the fileType, "nifti-single", "nifti-pair", or "nifti-gz"}
}
\value{
an instance of class "NIFTIFile"
}
\examples{
}

\keyword{ manip }

