\name{NIfTIFileDescriptor-class}
\alias{NIfTIFileDescriptor-class}
\alias{NIfTIFileDescriptor}
\title{This class supports the NIfTI file format...}
\description{This class supports the NIfTI file format}
\section{Extends}{\code{\linkS4class{BrainFileDescriptor}}}
\section{Methods}{\describe{\item{\code{\link[=readMetaInfo,NIfTIFileDescriptor-method]{readMetaInfo}}}{}}}

