\name{NIfTIMetaInfo}
\alias{NIfTIMetaInfo}
\title{Constructor for \linkS4class{NIfTIMetaInfo} class...}
\usage{NIfTIMetaInfo(descriptor, nifti_header)
}
\description{Constructor for \code{\linkS4class{NIfTIMetaInfo}} class}
\value{an instance of class \code{\linkS4class{NIfTIMetaInfo}}}
\arguments{\item{descriptor}{an instance of class \code{\linkS4class{NIfTIFileDescriptor}}}
\item{nifti_header}{a \code{list} returned by \code{readNIftiHeader}}
}

