% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{NIMLSurfaceDataMetaInfo-class}
\alias{NIMLSurfaceDataMetaInfo-class}
\title{NIMLSurfaceDataMetaInfo
	
This class contains meta information for surface-based data for the NIML data format}
\description{
NIMLSurfaceDataMetaInfo
	
This class contains meta information for surface-based data for the NIML data format
}
\section{Slots}{

\describe{
\item{\code{data}}{the numeric data matrix of surface values (rows = nodes, columns=surface vectors)}

\item{\code{nodeIndices}}{the indices of the nodes for mapping to associated surface geometry.}
}}

