% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainMetaInfo.R
\name{NIMLSurfaceDataMetaInfo}
\alias{NIMLSurfaceDataMetaInfo}
\title{Constructor for \code{\linkS4class{NIMLSurfaceDataMetaInfo}} class}
\usage{
NIMLSurfaceDataMetaInfo(descriptor, header)
}
\arguments{
\item{descriptor}{the file descriptor}

\item{header}{a \code{list} containing header information}
}
\description{
Constructor for \code{\linkS4class{NIMLSurfaceDataMetaInfo}} class
}

