\name{NullMetaInfo-class}
\alias{NullMetaInfo-class}
\alias{NullMetaInfo}
\title{This is class is used to denote the absense of meta information...}
\description{This is class is used to denote the absense of meta information}
\section{Extends}{\code{\linkS4class{BaseMetaInfo}}}
\section{Methods}{\describe{\item{\code{\link[=show,NullMetaInfo-method]{show}}}{}}}

