\name{numdim}
\alias{numdim}
\title{ the number of dimensions in image }
\description{
  generic function to access the number of dimensions (e.g. axes) in a
  brain iamge
}
\usage{
numdim(x)
}

\arguments{
  \item{x}{ an object derived from "BrainSpace" or "BrainData" class }
  
}

\value{
an integer value for the number of dimensions
}
\examples{

}

\keyword{ manip }

