\name{openFor}
\alias{openFor}
\title{ indicates whether the file is open for read or write  }
\description{
  function which indicates whether a "BrainFile" object is open for read
  or write operations
}
\usage{
openFor(x)
}

\arguments{
  \item{x}{ an object derived from "BrainFile" class }
}

\value{
  returns "r" for read and "w" for write
}
\examples{
}

\keyword{ manip }

