\name{origin}
\alias{origin}

\title{ retrieves the image origin }
\description{

  the image is the real world coordinate location at the first grid
  location of the data array (e.g. voxel [1,1,1] )


}
\usage{
origin(x)
}

\arguments{
  \item{x}{ object deriving from class "BrainSpace" or "BrainData" }
}

\value{
  a numeric vector indicating the location in real world coordinates.
}

\examples{}

\keyword{ manip }

