% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{patchSet}
\alias{patchSet}
\title{patchSet}
\usage{
patchSet(x, dims, mask, ...)
}
\arguments{
\item{x}{the object to extract patches from}

\item{dims}{a vector indicating the dimensions of the patches}

\item{mask}{mask indicating the valid patch area}
}
\description{
patchSet
}
