\name{path}
\alias{path}

\title{ returns the full path of a "BrainFile" source  }
\description{
  returns the full path of a "BrainFile" source 
}
\usage{
path(x)
}


\arguments{
  \item{x}{ an object derived from the "BrainFile" class  }
}
\value{
 character string representing the full path of the data file
}
\examples{

}

\keyword{ manip }

