% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Axis.R
\docType{methods}
\name{permMat}
\alias{permMat}
\alias{permMat,AxisSet2D-method}
\alias{permMat,AxisSet3D-method}
\alias{permMat,BrainSpace-method}
\title{Extract permutation matrix}
\usage{
permMat(x, ...)

\S4method{permMat}{AxisSet2D}(x, ...)

\S4method{permMat}{AxisSet3D}(x, ...)

\S4method{permMat}{BrainSpace}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional arguments}
}
\description{
Extract permutation matrix

permMat

permMat

permMat
}
