\name{pick-methods}
\docType{methods}
\alias{pick-methods}
\alias{pick,BrainVector,vector-method}
\title{ extracts data from an image using a mask proivded as an argument }
\description{
 a generic function for "picking" or extracting data from an image based on a mask. 
}
\section{Methods}{
\describe{

\item{x = "BrainVector", mask = "vector"}{ extracts a matrix from a BrainVector instance using the supplied \code{mask} }
}}
\keyword{methods}

