% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Axis.R
\docType{methods}
\name{print,AxisSet2D-method}
\alias{print,AxisSet2D-method}
\title{print a \code{AxisSet2D} instance}
\usage{
\S4method{print}{AxisSet2D}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{extra args}
}
\description{
print a \code{AxisSet2D} instance
}

