% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Axis.R
\docType{methods}
\name{print,AxisSet3D-method}
\alias{print,AxisSet3D-method}
\title{print a \code{AxisSet3D} instance}
\usage{
\S4method{print}{AxisSet3D}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{extra args}
}
\description{
print a \code{AxisSet3D} instance
}

