\docType{methods}
\name{print,NamedAxis-method}
\alias{print,AxisSet2D-method}
\alias{print,AxisSet3D-method}
\alias{print,NamedAxis-method}
\title{print a \code{NamedAxis} instance}
\usage{
\S4method{print}{NamedAxis}(x, ...)

\S4method{print}{AxisSet2D}(x, ...)

\S4method{print}{AxisSet3D}(x, ...)
}
\arguments{
  \item{x}{an object used to select a method.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\description{
print a \code{NamedAxis} instance

print a \code{AxisSet2D} instance

print a \code{AxisSet3D} instance
}

