% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Axis.R
\docType{methods}
\name{print,NamedAxis-method}
\alias{print,NamedAxis-method}
\title{print a \code{NamedAxis}}
\usage{
\S4method{print}{NamedAxis}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{extra arguments}
}
\description{
print a \code{NamedAxis}
}
