\name{private}
\alias{.BrainVectorFromIndices}
\title{.BrainVectorFromIndices}
\usage{.BrainVectorFromIndices(data, space, indices)
.BrainVectorFromMatrix(data, space)
.createVectorSpaceFromData(data, space, indices)
.pruneCoords(coord.set, vals, mindist=10)
.isExtension(fname, extension)
.concat4D(x, y, ...)
.extract.array(x, ..., drop=FALSE, indices=list(...))
.gridToIndex(dimensions, vmat)
.indexToGrid(idx, array.dim)
.getRStorage(dataType)
.getDataStorage(code)
.getDataCode(dataType)
.getDataSize(dataType)
.getEndian(conn)
.niftiExt(filetype)
.matrixToQuatern(mat)
.quaternToMatrix(quat, origin, stepSize, qfac)
}
\alias{.BrainVectorFromMatrix}
\alias{.createVectorSpaceFromData}
\alias{.pruneCoords}
\alias{.isExtension}
\alias{.concat4D}
\alias{.extract.array}
\alias{.gridToIndex}
\alias{.indexToGrid}
\alias{.getRStorage}
\alias{.getDataStorage}
\alias{.getDataCode}
\alias{.getDataSize}
\alias{.getEndian}
\alias{.niftiExt}
\alias{.matrixToQuatern}
\alias{.quaternToMatrix}

