% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainRegion3D.R
\name{RandomSurfaceSearchlight}
\alias{RandomSurfaceSearchlight}
\title{Create a Random Searchlight iterator for surface mesh using geodesic distance to define regions.}
\usage{
RandomSurfaceSearchlight(surfgeom, radius = 8, nodeset = NULL)
}
\arguments{
\item{surfgeom}{a surface mesh: instance of class \code{\linkS4class{SurfaceGeometry}}}

\item{radius}{radius of the searchlight as a geodesic distance in mm}

\item{nodeset}{the subset of surface node indices to use}
}
\description{
Create a Random Searchlight iterator for surface mesh using geodesic distance to define regions.
}
\details{
On every call to \code{nextElem} a set of surface nodes are returned. 
  These nodes index into the vertices of the \code{igraph} instance.
}
\examples{
file <- system.file("extdata", "std.lh.smoothwm.asc", package = "neuroim")
geom <- loadSurface(file)
searchlight <- RandomSurfaceSearchlight(geom, 12)
nodes <- searchlight$nextElem()
length(nodes) > 1

}

