% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AFNI_IO.R
\name{readAFNIHeader}
\alias{readAFNIHeader}
\title{readAFNIHeader}
\usage{
readAFNIHeader(fileName)
}
\arguments{
\item{fileName}{the name of the AFNI header file (ending in .HEAD)}
}
\description{
readAFNIHeader
}

