% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFNI_IO.R
\name{readAFNISurfaceHeader}
\alias{readAFNISurfaceHeader}
\title{readAFNISurfaceHeader}
\usage{
readAFNISurfaceHeader(fileName)
}
\arguments{
\item{fileName}{the name of the AFNI 1D file}
}
\description{
readAFNISurfaceHeader
}

