% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainMetaInfo.R
\docType{methods}
\name{readColumns}
\alias{readColumns}
\alias{readColumns,ColumnReader,numeric-method}
\title{Generic function to read a set of column vector from an input source (e.g. \code{ColumnReader})}
\usage{
readColumns(x, columnIndices)

\S4method{readColumns}{ColumnReader,numeric}(x, columnIndices)
}
\arguments{
\item{x}{the input channel}

\item{columnIndices}{the column indices}
}
\value{
a \code{matrix} consisting of the requested column vectors
}
\description{
Generic function to read a set of column vector from an input source (e.g. \code{ColumnReader})
}

