\name{readData}
\alias{readData}
\title{ reads image data into an appropriate "BrainData" object  }
\description{
  function to load image data into appropriate data structure, e.g. an
  object derived from the "BrainData" class
}
\usage{
readData(x, indices)
}

\arguments{
  \item{x}{ an object derived from the "BrainFile" class }
  \item{indices}{ the 1D indices of the image cells to load }
}

\details{
  
}
\value{
  a "BrainData" object appropriate for the type of data file being read
}

\examples{

}

\keyword{ manip }

