% Generated by roxygen2 (4.0.0): do not edit by hand
\name{readHeader}
\alias{readHeader}
\title{read header information of an image file}
\usage{
readHeader(fileName)
}
\arguments{
\item{fileName}{the name of the file to read}
}
\value{
an instance of class \code{\linkS4class{FileMetaInfo}}
}
\description{
read header information of an image file
}

