% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{readMetaInfo}
\alias{readMetaInfo}
\alias{readMetaInfo,AFNIFileDescriptor-method}
\alias{readMetaInfo,NIfTIFileDescriptor-method}
\title{Generic function to read image meta info given a file and a \code{\linkS4class{BrainFileDescriptor}} instance.}
\usage{
readMetaInfo(x, fileName)

\S4method{readMetaInfo}{NIfTIFileDescriptor}(x, fileName)

\S4method{readMetaInfo}{AFNIFileDescriptor}(x, fileName)
}
\arguments{
\item{x}{descriptor instance}

\item{fileName}{file name contianing meta information}
}
\description{
Generic function to read image meta info given a file and a \code{\linkS4class{BrainFileDescriptor}} instance.
}

