\name{readNIFTIHeader}
\alias{readNIFTIHeader}

\title{ reads a NIFTI format header }
\description{
  function to read a NIFTI formatted header file 
}
\usage{
readNIFTIHeader(fname)
}

\arguments{
  \item{fname}{ the file name of the header }
}

\value{
  an object of class "NIFTIHeader"
}

\examples{
}

\keyword{ manip }

