% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFNI_IO.R
\name{readNIMLSurfaceHeader}
\alias{readNIMLSurfaceHeader}
\title{readNIMLSurfaceHeader}
\usage{
readNIMLSurfaceHeader(fileName)
}
\arguments{
\item{fileName}{the name of the NIML file}
}
\description{
readNIMLSurfaceHeader
}

