% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainRegion3D.R
\name{RegionSphere}
\alias{RegionSphere}
\title{Create a Spherical Region of Interest}
\usage{
RegionSphere(bvol, centroid, radius, fill = NULL, nonzero = FALSE)
}
\arguments{
\item{bvol}{an \code{BrainVolume} or \code{BrainSpace} instance}

\item{centroid}{the center of the sphere in voxel space}

\item{radius}{the radius in real units (e.g. millimeters) of the spherical ROI}

\item{fill}{optional value(s) to store as data}

\item{nonzero}{if \code{TRUE}, keep only nonzero elements from \code{bvol}}
}
\value{
an instance of class \code{ROIVolume}
}
\description{
Creates a Spherical ROI based on a Centroid.
}
\examples{
 sp1 <- BrainSpace(c(10,10,10), c(1,2,3))
 cube <- RegionSphere(sp1, c(5,5,5), 3.5)
 vox <- coords(cube)
 cds <- coords(cube, real=TRUE)
 ## fill in ROI with value of 6
 cube1 <- RegionSphere(sp1, c(5,5,5), 3.5, fill=6)
 all(cube1@data == 6)
}
