% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Display.R
\docType{methods}
\name{render}
\alias{render}
\alias{render,BrainSlice,numeric,numeric,character-method}
\title{Render an image to create a drawable image.}
\usage{
render(x, width, height, colmap, ...)

\S4method{render}{BrainSlice,numeric,numeric,character}(x, width, height,
  colmap, zero.col = "#000000FF", alpha = 1, units = "mm")
}
\arguments{
\item{x}{the object, e.g. an instance of type \code{BrainSlice}}

\item{width}{width of the rendered image}

\item{height}{height of the rendered image}

\item{colmap}{the colors used to map from values to RGBA colors.}

\item{...}{additional arguments}

\item{zero.col}{color used when background intensity is 0.}

\item{alpha}{transparency multiplier}

\item{units}{grid unit type, e.g. "mm", "inches"}
}
\description{
Render an image to create a drawable image.
}

