% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{reorient}
\alias{reorient}
\title{Generic function to remap the grid-to-world coordinates mapping of an image.}
\usage{
reorient(x, orient)
}
\arguments{
\item{x}{the object}

\item{orient}{the orientation code indcating the "remapped" axes.}
}
\value{
a reoriented space
}
\description{
Generic function to remap the grid-to-world coordinates mapping of an image.
}
