% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ROISurface-class}
\alias{ROISurface-class}
\title{ROISurface}
\description{
A class that respresents a surface-based region of interest
}
\section{Slots}{

\describe{
\item{\code{geometry}}{the geometry of the parent surface: a \code{SurfaceGeometry} instance}

\item{\code{data}}{the \code{numeric} or \code{matrix} data stored in ROI}

\item{\code{coords}}{the surface-based coordinates of the data}

\item{\code{indices}}{the node indices of the parent surface stored in the \code{geometry} field.}
}}

