% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainRegion3D.R
\name{ROISurface}
\alias{ROISurface}
\title{Create an instance of class \code{\linkS4class{ROISurface}}}
\usage{
ROISurface(geometry, indices, data)
}
\arguments{
\item{geometry}{the parent geometry: an instance of class \code{SurfaceGeometry}}

\item{indices}{the parent surface indices}

\item{data}{the data values, numeric \code{vector}}
}
\value{
an instance of class \code{ROISurface}
}
\description{
Create an instance of class \code{\linkS4class{ROISurface}}
}

