% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainRegion3D.R
\name{ROISurfaceVector}
\alias{ROISurfaceVector}
\title{Create an instance of class \code{\linkS4class{ROISurfaceVector}}}
\usage{
ROISurfaceVector(geometry, indices, data)
}
\arguments{
\item{geometry}{the parent geometry: an instance of class \code{SurfaceGeometry}}

\item{indices}{the parent surface indices}

\item{data}{the data values, a \code{matrix}}
}
\value{
an instance of class \code{ROISurfaceVector}
}
\description{
Create an instance of class \code{\linkS4class{ROISurfaceVector}}
}

