% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ROIVector-class}
\alias{ROIVector-class}
\title{ROIVector}
\description{
A class that represents a vector-valued volumetric region of interest
}
\section{Slots}{

\describe{
\item{\code{data}}{the \code{matrix} data stored in ROI}

\item{\code{coords}}{the voxel coordinates of the ROI}
}}

