% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainRegion3D.R
\name{ROIVector}
\alias{ROIVector}
\title{Create an instance of class \code{\linkS4class{ROIVector}}}
\usage{
ROIVector(vspace, coords, data = rep(nrow(coords), 1))
}
\arguments{
\item{vspace}{an instance of class \code{BrainSpace}}

\item{coords}{matrix of voxel coordinates}

\item{data}{the \code{matrix} of data values}
}
\value{
an instance of class \code{ROIVector}
}
\description{
Create an instance of class \code{\linkS4class{ROIVector}}
}
