\docType{class}
\name{ROIVolume-class}
\alias{ROIVolume-class}
\title{ROIVolume}
\description{
A class that is used to produce a
\code{\linkS4class{SparseBrainVector}} instance
}
\section{Slots}{

\describe{
\item{\code{data}}{the data stored in the ROI}

\item{\code{coords}}{the coordinates of the ROI}
}}

