\name{Searchlight}
\alias{Searchlight}
\title{Create an exhaustive searchlight iterator}
\usage{
Searchlight(mask, radius)
}
\arguments{
  \item{mask}{an image volume containing valid central
  voxels for roving searchlight}

  \item{radius}{in mm of spherical searchlight}
}
\description{
Create an exhaustive searchlight iterator
}

