% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainVector.R, R/SparseBrainVector.R
\docType{methods}
\name{seriesIter}
\alias{seriesIter}
\alias{seriesIter,BrainVector-method}
\alias{seriesIter,SparseBrainVector-method}
\title{Construct a series iterator}
\usage{
seriesIter(x)

\S4method{seriesIter}{BrainVector}(x)

\S4method{seriesIter}{SparseBrainVector}(x)
}
\arguments{
\item{x}{the object to be iterated}
}
\description{
Construct a series iterator
}

