% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Axis.R
\docType{methods}
\name{show,AxisSet1D-method}
\alias{show,AxisSet1D-method}
\title{show an \code{AxisSet1D}}
\usage{
\S4method{show}{AxisSet1D}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show an \code{AxisSet1D}
}
