% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Axis.R
\docType{methods}
\name{show,AxisSet2D-method}
\alias{show,AxisSet2D-method}
\title{show an \code{AxisSet2D}}
\usage{
\S4method{show}{AxisSet2D}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show an \code{AxisSet2D}
}

