% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Axis.R
\docType{methods}
\name{show,AxisSet4D-method}
\alias{show,AxisSet4D-method}
\title{show an \code{AxisSet4D}}
\usage{
\S4method{show}{AxisSet4D}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show an \code{AxisSet4D}
}

