% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{show,BaseMetaInfo-method}
\alias{show,BaseMetaInfo-method}
\title{show a \code{BaseMetaInfo}}
\usage{
\S4method{show}{BaseMetaInfo}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show a \code{BaseMetaInfo}
}
