% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\docType{methods}
\name{show,BrainSurfaceVector-method}
\alias{show,BrainSurfaceVector-method}
\title{show a \code{BrainSurfaceVector}}
\usage{
\S4method{show}{BrainSurfaceVector}(object)
}
\arguments{
\item{x}{the object}

\item{...}{extra arguments}
}
\description{
show a \code{BrainSurfaceVector}
}

