% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrainVector.R
\docType{methods}
\name{show,BrainVectorSource-method}
\alias{show,BrainVectorSource-method}
\title{show a \code{BrainVectorSource}}
\usage{
\S4method{show}{BrainVectorSource}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show a \code{BrainVectorSource}
}

