% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrainVolume.R
\docType{methods}
\name{show,BrainVolume-method}
\alias{show,BrainVolume-method}
\title{show a \code{BrainVolume}}
\usage{
\S4method{show}{BrainVolume}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show a \code{BrainVolume}
}

