% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrainMetaInfo.R
\docType{methods}
\name{show,FileMetaInfo-method}
\alias{show,FileMetaInfo-method}
\title{show a \code{FileMetaInfo}}
\usage{
\S4method{show}{FileMetaInfo}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show a \code{FileMetaInfo}
}

