% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrainRegion3D.R, R/BrainVector.R, R/BrainVolume.R
\docType{methods}
\name{show,ROIVolume-method}
\alias{show,BrainVector-method}
\alias{show,BrainVectorSource-method}
\alias{show,BrainVolume-method}
\alias{show,ROIVolume-method}
\title{show}
\usage{
\S4method{show}{ROIVolume}(object)

\S4method{show}{BrainVectorSource}(object)

\S4method{show}{BrainVector}(object)

\S4method{show}{BrainVolume}(object)
}
\arguments{
\item{object}{the object to \code{show}}

\item{the}{object to \code{show}}
}
\description{
show

show
}

