% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Axis.R
\docType{methods}
\name{show,NamedAxis-method}
\alias{show,NamedAxis-method}
\title{show an \code{NamedAxis}}
\usage{
\S4method{show}{NamedAxis}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show an \code{NamedAxis}
}

