% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{show,NullMetaInfo-method}
\alias{show,NullMetaInfo-method}
\title{show a \code{NullMetaInfo}}
\usage{
\S4method{show}{NullMetaInfo}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show a \code{NullMetaInfo}
}
