% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrainRegion3D.R
\docType{methods}
\name{show,ROIVolume-method}
\alias{show,ROIVolume-method}
\title{show an \code{ROIVolime}}
\usage{
\S4method{show}{ROIVolume}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show an \code{ROIVolime}
}

