% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseBrainVector.R
\docType{methods}
\name{show,SparseBrainVector-method}
\alias{show,SparseBrainVector-method}
\title{show a \code{SparseBrainVector}}
\usage{
\S4method{show}{SparseBrainVector}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show a \code{SparseBrainVector}
}

