% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainMetaInfo.R
\docType{methods}
\name{show,SurfaceDataMetaInfo-method}
\alias{show,SurfaceDataMetaInfo-method}
\title{show an \code{SurfaceDataMetaInfo}}
\usage{
\S4method{show}{SurfaceDataMetaInfo}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show an \code{SurfaceDataMetaInfo}
}

