% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainMetaInfo.R
\docType{methods}
\name{show,SurfaceGeometryMetaInfo-method}
\alias{show,SurfaceGeometryMetaInfo-method}
\title{show an \code{SurfaceGeometryMetaInfo}}
\usage{
\S4method{show}{SurfaceGeometryMetaInfo}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show an \code{SurfaceGeometryMetaInfo}
}

