% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{sliceData}
\alias{sliceData}
\title{sliceData
extract a 2D slice from a \code{BrainVolume} instance.}
\usage{
sliceData(vol, slice, axis = 3)
}
\arguments{
\item{vol}{an \code{BrainVolume} instance}

\item{slice}{the integer index of the slice to cut.}

\item{axis}{the axis number (1, 2, 3) defining fixed axis of the 2D slice.}
}
\description{
sliceData
extract a 2D slice from a \code{BrainVolume} instance.
}

