\name{sliceMeans-methods}
\docType{methods}
\alias{sliceMeans-methods}
\alias{sliceMeans,BrainVector-method}
\alias{sliceMeans}
\title{ function that computes the mean over slices in the 4th (time) dimension }
\description{
 Methods for function \code{sliceMeans}  
}
\section{Methods}{
\describe{

\item{x = "BrainVector"}{ ~~describe this method here }
}}
\keyword{methods}

