% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainData.R, R/BrainSpace.R, R/IndexLookupVolume.R
\docType{methods}
\name{space}
\alias{space}
\alias{space,BrainData-method}
\alias{space,BrainSpace-method}
\alias{space,IndexLookupVolume-method}
\title{Generic function to extract the \code{space} member variable}
\usage{
space(x, ...)

\S4method{space}{BrainData}(x)

\S4method{space}{BrainSpace}(x)

\S4method{space}{IndexLookupVolume}(x)
}
\arguments{
\item{x}{the object to query}

\item{...}{additional arguments}
}
\value{
an object representing the geometric space of the image
}
\description{
Generic function to extract the \code{space} member variable

space

space

space
}

