\name{space}
\alias{space}
\title{ accessor for "BrainSpace" slot }
\description{
  generic function to access the spatial geometry of a brain image 
}
\usage{
space(x)
}

\arguments{
  \item{x}{ an object derived the "BrainData" class }
  
}

\value{
an object of class \code{BrainSpace}
}
\examples{


}

\keyword{ manip }

