% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainRegion3D.R
\docType{methods}
\name{[,ROISurface,logical,missing,ANY-method}
\alias{[,ROISurface,logical,missing,ANY-method}
\title{subset an \code{ROISurface}}
\usage{
\S4method{[}{ROISurface,logical,missing,ANY}(x, i, j, drop)
}
\arguments{
\item{x}{the object}

\item{i}{first index}

\item{j}{second index}

\item{drop}{drop dimension}
}
\description{
subset an \code{ROISurface}
}

