% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrainRegion3D.R
\docType{methods}
\name{[,ROIVolume,numeric,missing,ANY-method}
\alias{[,ROIVolume,numeric,missing,ANY-method}
\title{extract data from \code{ROIVolume}}
\usage{
\S4method{[}{ROIVolume,numeric,missing,ANY}(x, i, j, drop)
}
\arguments{
\item{j}{index for second dimension (missing)}
}
\description{
extract data from \code{ROIVolume}
}

