% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrainVolume.R
\docType{methods}
\name{[,SparseBrainVolume,missing,missing,ANY-method}
\alias{[,SparseBrainVolume,missing,missing,ANY-method}
\title{extractor}
\usage{
\S4method{[}{SparseBrainVolume,missing,missing,ANY}(x, i, j, k, ...,
  drop = TRUE)
}
\arguments{
\item{x}{the object}

\item{i}{first index}

\item{j}{second index}

\item{k}{third index}

\item{...}{additional args}

\item{drop}{dimension}
}
\description{
extractor
}

