% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainVolume.R
\docType{methods}
\name{[,SparseBrainVolume,numeric,missing,missing-method}
\alias{[,SparseBrainVolume,numeric,missing,missing-method}
\title{extractor}
\usage{
\S4method{[}{SparseBrainVolume,numeric,missing,missing}(x, i, j, k, ..., drop)
}
\arguments{
\item{x}{the object}

\item{i}{first index}

\item{j}{second index}

\item{k}{third index}

\item{...}{additional args}

\item{drop}{drop dimension}
}
\description{
extractor
}
