% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BrainVector.R
\docType{methods}
\name{[[,BrainBucket,character,missing-method}
\alias{[[,BrainBucket,character,missing-method}
\title{extract labeled volume from \code{BrainBucket}}
\usage{
\S4method{[[}{BrainBucket,character,missing}(x, i)
}
\description{
extract labeled volume from \code{BrainBucket}
}

