% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainBucket.R
\docType{methods}
\name{[[,BrainBucket,index,missing-method}
\alias{[[,BrainBucket,index,missing-method}
\title{extract labeled volume from \code{BrainBucket}}
\usage{
\S4method{[[}{BrainBucket,index,missing}(x, i)
}
\arguments{
\item{x}{the object}

\item{i}{the first index}
}
\description{
extract labeled volume from \code{BrainBucket}
}
