% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BrainVector.R
\docType{methods}
\name{[[,BrainBucket,numeric,missing-method}
\alias{[[,BrainBucket,numeric,missing-method}
\title{extract indexed volume from \code{BrainBucket}}
\usage{
\S4method{[[}{BrainBucket,numeric,missing}(x, i)
}
\description{
extract indexed volume from \code{BrainBucket}
}

