% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainRegion3D.R
\name{SurfaceDisk}
\alias{SurfaceDisk}
\title{Create a Region on Surface}
\usage{
SurfaceDisk(surf, index, radius, max_order = NULL)
}
\arguments{
\item{surf}{a \code{SurfaceGeometry} or \code{BrainSurface} or \code{BrainSurfaceVector}}

\item{index}{the index of the central surface node}

\item{radius}{the size in mm of the geodesic radius}

\item{max_order}{maximum number of edges to traverse. 
default is computed based on average edge length.}
}
\description{
Creates a Region on a Surface from a radius and surface
}

