% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SurfaceGeometryMetaInfo-class}
\alias{SurfaceGeometryMetaInfo-class}
\title{SurfaceGeometryMetaInfo
	
This class contains meta information for brain surface geometry}
\description{
SurfaceGeometryMetaInfo
	
This class contains meta information for brain surface geometry
}
\section{Slots}{

\describe{
\item{\code{nodes}}{the number of surface vertices}

\item{\code{faces}}{the number of faces}

\item{\code{embedDimension}}{the dimensionality of the embedding}

\item{\code{label}}{a label indicating the type of surface (e.g. white, pial, inflated, flat, spherical)}
}}

