% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R, R/BrainSurface.R
\docType{class}
\name{SurfaceGeometrySource-class}
\alias{SurfaceGeometrySource}
\alias{SurfaceGeometrySource-class}
\title{SurfaceGeometrySource}
\usage{
SurfaceGeometrySource(surfaceName)
}
\arguments{
\item{surfaceName}{the name of the file containing the surface geometry.}
}
\description{
A class that is used to produce a \code{\linkS4class{SurfaceGeometry}} instance

Constructor for SurfaceGeometrySource
}
\section{Slots}{

\describe{
\item{\code{metaInfo}}{a \code{\linkS4class{SurfaceGeometryMetaInfo}} instance}
}}

