% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SurfaceMetaInfo-class}
\alias{SurfaceMetaInfo-class}
\title{SurfaceMetaInfo}
\description{
This class contains meta information for a brain surface
}
\section{Slots}{

\describe{
\item{\code{nodes}}{the number of surface vertices}

\item{\code{faces}}{the number of faces}

\item{\code{embedDimension}}{the dimensionality of the embedding}

\item{\code{type}}{a label indicating the type of surface (e.g. white, pial, inflated, flat, spherical)}
}}

