% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainRegion3D.R
\name{SurfaceSearchlight}
\alias{SurfaceSearchlight}
\title{Create a Searchlight iterator for surface mesh using geodesic distance to define regions.}
\usage{
SurfaceSearchlight(surfgeom, radius = 8, nodeset = NULL)
}
\arguments{
\item{surfgeom}{a surface mesh: instance of class \code{SurfaceGeometry}}

\item{radius}{radius of the searchlight as a geodesic distance in mm}

\item{nodeset}{the subset of surface nodes to use}
}
\description{
Create a Searchlight iterator for surface mesh using geodesic distance to define regions.
}

