\name{takeSeries}
\alias{takeSeries}
\title{Generic function to extract a set of series from a 4D image}
\usage{
takeSeries(x, indices, ...)
}
\arguments{
  \item{x}{a four dimensional image}

  \item{i}{the indices of the series' to extract}
}
\description{
Generic function to extract a set of series from a 4D image
}

