% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainVector.R, R/SparseBrainVector.R
\docType{methods}
\name{takeVolume}
\alias{takeVolume}
\alias{takeVolume,BrainVector,numeric-method}
\alias{takeVolume,SparseBrainVector,numeric-method}
\title{Generic function to extract a volume from a four-dimensional image}
\usage{
takeVolume(x, i, ...)

\S4method{takeVolume}{BrainVector,numeric}(x, i, merge = FALSE)

\S4method{takeVolume}{SparseBrainVector,numeric}(x, i, merge = FALSE)
}
\arguments{
\item{x}{four-dimensional image}

\item{i}{the indices of the volume(s) to extract}

\item{...}{additional arguments}

\item{merge}{concatenate extracted volumes}
}
\description{
Generic function to extract a volume from a four-dimensional image

takeVolume
extract volume from SparseBrainVector
}

