\name{versionNum}
\alias{versionNum}
\title{ retrieves the version number associated with an image file format }
\description{
  retrieves the version number associated with an image file format 
}
\usage{
versionNum(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ an object derived from class "BrainHeader" }
}

\value{
  a character string indicating the version number
}

\examples{
}

\keyword{ manip }

