% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainRegion3D.R
\docType{methods}
\name{voxels}
\alias{voxels}
\alias{voxels,Kernel-method}
\title{extract voxel coordinates}
\usage{
voxels(x, ...)

\S4method{voxels}{Kernel}(x, centerVoxel = NULL)
}
\arguments{
\item{x}{the object to extract voxels from}

\item{...}{additional arguments to function}

\item{centerVoxel}{the absolute location of the center of the voxel, default is (0,0,0)}
}
\description{
extract voxel coordinates

extract voxels from a \code{Kernel} object
}

