\name{writeData}

\alias{writeData}

\title{ write image data to disk }

\description{
  save image data stored in memory to the file system 
}
\usage{
writeData(x, header, data)
}

\arguments{
  \item{x}{ an object deriving from class "BrainFile" }
  \item{header}{ the header object assoicated with data }
  \item{data}{ the raw numeric data to write }
}

\value{
  no return value
}


\keyword{ manip }

