\name{writeHeader}
\alias{writeHeader}
\title{ writes image header to disk }
\description{
  generic function to store contents of image header to disk.
}
\usage{
writeHeader(x, header)
}

\arguments{
  \item{x}{ obect deriving from class "BrainFile" }
  \item{header}{ object deriving from class "BrainHeader" }
}

\value{
  no return value
}


\examples{

}

\keyword{manip }
