\name{writeVolume}
\alias{writeVolume}

\title{ function to write a three-dimensional image volume to disk }
\description{
  function to write a three-dimensional image volume to disk
}
\usage{
writeVolume(x, fileName, format)
}

\arguments{
  \item{x}{ image volume to write }
  \item{fileName}{ name of file to write to }
  \item{format}{ image output format }
}

\examples{
}

\keyword{ manip }

