% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{addDim}
\alias{addDim}
\alias{addDim,BrainSpace,numeric-method}
\title{Generic function to add a dimension to an object}
\usage{
addDim(x, n)

\S4method{addDim}{BrainSpace,numeric}(x, n)
}
\arguments{
\item{x}{a dimensioned object}

\item{n}{the size of the dimension to add}
}
\description{
Generic function to add a dimension to an object

add dimension to \code{\linkS4class{BrainSpace}}
}

