\name{addDim-methods}
\alias{addDim-methods}
\alias{addDim}
\title{Generic function to add a dimension to an object...}
\description{Generic function to add a dimension to an object}
\arguments{\item{x}{a dimensioned object}
\item{n}{the size of the dimension to add}
}
\alias{addDim,BrainSpace,numeric-method}
\alias{addDim}
\details{\code{addDim,BrainSpace,numeric-method}: function is used to add a dimension to a \code{\linkS4class{BrainSpace}} object

}
\alias{dropDim,BrainSpace-method}
\alias{dropDim}

