% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{AFNIMetaInfo}
\alias{AFNIMetaInfo}
\title{AFNIMetaInfo}
\usage{
AFNIMetaInfo(descriptor, afni_header)
}
\arguments{
\item{descriptor}{an instance of class \code{\linkS4class{AFNIFileDescriptor}}}

\item{afni_header}{a \code{list} returned by \code{readAFNIHeader}}
}
\value{
an instance of class \code{\linkS4class{AFNIMetaInfo}}
}
\description{
Constructor for \code{\linkS4class{AFNIMetaInfo}} class
}

