\name{AFNIMetaInfo-class}
\alias{AFNIMetaInfo-class}
\alias{AFNIMetaInfo}
\title{This class contains meta information for a AFNI image file...}
\description{This class contains meta information for a AFNI image file}
\section{Slots}{\describe{\item{\code{afni_header}:}{}}}
\section{Extends}{\code{\linkS4class{BrainMetaInfo}}}

