\name{AFNIMetaInfo-class}
\alias{AFNIMetaInfo-class}
\alias{AFNIMetaInfo}
\title{This class contains meta information for a AFNI image file...}
\description{This class contains meta information for a AFNI image file}
\section{Slots}{\describe{\item{\code{afni_header}:}{}}}
\section{Extends}{\code{\linkS4class{FileMetaInfo}}}
\section{Methods}{\describe{\item{\code{\link[=dataReader,AFNIMetaInfo-method]{dataReader}}}{}}}

