% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainRegion3D.R, R/BrainVolume.R
\name{as}
\alias{as}
\title{conversion from DenseBrainVolume to array}
\description{
conversion from DenseBrainVolume to array

conversion from \code{SparseBrainVolume} to \code{array}

conversion from SparseBrainVolume to numeric

conversion from BrainVolume to LogicalBrainVolume

conversion from DenseBrainVolume to LogicalBrainVolume

conversion from ClusteredBrainVolume to LogicalBrainVolume

conversion from BrainVolume to array
}
