% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{as.list,BrainVector-method}
\alias{as.list,BrainVector-method}
\alias{as.list,SparseBrainVector-method}
\title{convert a BrainVector to a \code{list} of volumes}
\usage{
\S4method{as.list}{BrainVector}(x)

\S4method{as.list}{SparseBrainVector}(x)
}
\arguments{
\item{x}{object to be coerced or tested.}
}
\description{
convert a BrainVector to a \code{list} of volumes

convert SparseBrainVector to list of DenseBrainVolumes
}

