% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrainVector.R, R/SparseBrainVector.R
\docType{methods}
\name{as.list,BrainVector-method}
\alias{as.list,BrainVector-method}
\alias{as.list,SparseBrainVector-method}
\title{convert a \code{BrainVector} to \code{list} of volumes.}
\usage{
\S4method{as.list}{BrainVector}(x)

\S4method{as.list}{SparseBrainVector}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert a \code{BrainVector} to \code{list} of volumes.

convert SparseBrainVector to list of \code{\linkS4class{DenseBrainVolume}}
}

