% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseBrainVector.R, R/BrainVector.R
\docType{methods}
\name{as.list,SparseBrainVector-method}
\alias{as.list,SparseBrainVector-method}
\alias{as.list,BrainVector-method}
\title{as.list}
\usage{
\S4method{as.list}{SparseBrainVector}(x)

\S4method{as.list}{BrainVector}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert SparseBrainVector to list of \code{\linkS4class{DenseBrainVolume}}
}
