% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BrainVolume.R
\docType{methods}
\name{as.logical,BrainVolume-method}
\alias{as.logical,BrainVolume-method}
\title{as.logical}
\usage{
\S4method{as.logical}{BrainVolume}(x)
}
\description{
Convert BrainVolume to logical vector
}

