% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainVolume.R
\docType{methods}
\name{as.logical,BrainVolume-method}
\alias{as.logical,BrainVolume-method}
\title{as.logical}
\usage{
\S4method{as.logical}{BrainVolume}(x)
}
\arguments{
\item{x}{the object}
}
\value{
an instance of \code{linkS4class{LogicalBrainVolume}}
}
\description{
Convert BrainVolume to \code{linkS4class{LogicalBrainVolume}}
}
\details{
the image values will be converted to using R base function \code{as.logical} and wrapped in \code{LogicalBrainVolume}
}

