% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrainData.R
\docType{methods}
\name{as.matrix,BrainData-method}
\alias{as.matrix,BrainData-method}
\title{convert \code{BrainData} instance to matrix}
\usage{
\S4method{as.matrix}{BrainData}(x)
}
\description{
convert \code{BrainData} instance to matrix
}

