% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{as.matrix,BrainData-method}
\alias{as.matrix,BrainData-method}
\title{convert \code{BrainData} instance to matrix}
\usage{
\S4method{as.matrix}{BrainData}(x)
}
\arguments{
\item{x}{an \R object.}
}
\description{
convert \code{BrainData} instance to matrix
}

