% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseBrainVector.R, R/BrainVector.R
\docType{methods}
\name{as.matrix,SparseBrainVector-method}
\alias{as.matrix,DenseBrainVector-method}
\alias{as.matrix,SparseBrainVector-method}
\title{as.matrix}
\usage{
\S4method{as.matrix}{SparseBrainVector}(x)

\S4method{as.matrix}{DenseBrainVector}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert SparseBrainVector to matrix

convert a \code{DenseBrainVector} to a matrix
}

