% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrainVector.R, R/SparseBrainVector.R
\docType{methods}
\name{as.matrix,DenseBrainVector-method}
\alias{as.matrix,DenseBrainVector-method}
\alias{as.matrix,SparseBrainVector-method}
\title{convert a \code{DenseBrainVector} to a matrix}
\usage{
\S4method{as.matrix}{DenseBrainVector}(x)

\S4method{as.matrix}{SparseBrainVector}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert a \code{DenseBrainVector} to a matrix

convert SparseBrainVector to matrix
}

