% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainRegion3D.R, R/SparseBrainVector.R,
%   R/BrainVector.R
\docType{methods}
\name{as.matrix,ROIVector-method}
\alias{as.matrix,ROIVector-method}
\alias{as.matrix,SparseBrainVector-method}
\alias{as.matrix,DenseBrainVector-method}
\title{convert a \code{ROIVector} to a matrix}
\usage{
\S4method{as.matrix}{ROIVector}(x)

\S4method{as.matrix}{ROIVector}(x)

\S4method{as.matrix}{SparseBrainVector}(x)

\S4method{as.matrix}{DenseBrainVector}(x)
}
\arguments{
\item{x}{the object}

\item{x}{the object}

\item{x}{the object}
}
\description{
convert SparseBrainVector to matrix
}
