% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainVolume.R
\docType{methods}
\name{as.numeric,SparseBrainVolume-method}
\alias{as.numeric,SparseBrainVolume-method}
\title{Convert SparseBrainVolume to numeric}
\usage{
\S4method{as.numeric}{SparseBrainVolume}(x)
}
\arguments{
\item{x}{the object to convert}
}
\description{
Convert SparseBrainVolume to numeric
}
