% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\docType{methods}
\name{as.raster,Layer-method}
\alias{as.raster,Layer-method}
\title{as.raster}
\usage{
\S4method{as.raster}{Layer}(x, zpos)
}
\arguments{
\item{x}{the layer to convert}

\item{zpos}{the z coordinate in coordinate space}
}
\description{
as.raster
}

