% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Display.R
\docType{methods}
\name{as.raster,Layer-method}
\alias{as.raster,Layer-method}
\title{as.raster}
\usage{
\S4method{as.raster}{Layer}(x, zpos, thresh = c(0, 0), axis = 3)
}
\arguments{
\item{x}{the layer to convert}

\item{zpos}{the z coordinate in coordinate space}

\item{thresh}{the threshold range}

\item{axis}{the axis index (1,2,3)}
}
\description{
as.raster
}

