% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainVector.R
\docType{methods}
\name{as.sparse}
\alias{as.sparse}
\alias{as.sparse,DenseBrainVector,numeric-method}
\title{Convert to sparse representation}
\usage{
as.sparse(x, mask, ...)

\S4method{as.sparse}{DenseBrainVector,numeric}(x, mask)
}
\arguments{
\item{x}{the object to sparsify}

\item{mask}{the elements to retain}

\item{...}{additional arguments}
}
\description{
Convert to sparse representation

convert a DenseBrainVector to a SparseBrainVector
}

