% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainSpace.R
\docType{methods}
\name{axes}
\alias{axes}
\alias{axes,BrainSpace-method}
\alias{axes,BrainData-method}
\title{Generic getter function to extract image axes}
\usage{
axes(x)

\S4method{axes}{BrainSpace}(x)

\S4method{axes}{BrainData}(x)
}
\arguments{
\item{x}{an object with a set of axes}
}
\description{
Generic getter function to extract image axes
}
