\name{AxisSet-class}
\alias{AxisSet-class}
\alias{AxisSet}
\title{Virtual base class representing an ordered set of named axes.}
\description{Virtual base class representing an ordered set of named axes.}
\section{Slots}{\describe{\item{\code{ndim}:}{}}}
\section{Methods}{\describe{\item{\code{\link[=ndim,AxisSet-method]{ndim}}}{\code{signature(x = “AxisSet”)}: get the number of dimensions for this AxisSet}}}

