\name{AxisSet-class}
\alias{AxisSet-class}
\alias{AxisSet}
\title{Virtual base class representing an ordered set of named axes.}
\description{Virtual base class representing an ordered set of named axes.}
\section{Slots}{\describe{\item{\code{ndim}:}{(\code{\link{integer}}) the number of axes (or dimensions)}}}
\section{Methods}{\describe{\item{\code{\link[=ndim,AxisSet-method]{ndim}}}{}}}

