\name{AxisSet2D-class}
\alias{AxisSet2D-class}
\alias{AxisSet2D}
\title{A two-dimensional axis set...}
\description{A two-dimensional axis set}
\section{Slots}{\describe{\item{\code{j}:}{(\code{\link{NamedAxis}}) the first axis}}}
\section{Extends}{\code{\linkS4class{AxisSet1D}}}
\section{Methods}{\describe{\item{\code{\link[=print,AxisSet2D-method]{print}}}{}\item{\code{\link[=show,AxisSet2D-method]{show}}}{}}}

