\name{AxisSet3D-class}
\alias{AxisSet3D-class}
\alias{AxisSet3D}
\title{A three-dimensional axis set...}
\description{A three-dimensional axis set}
\section{Slots}{\describe{\item{\code{k}:}{(\code{\link{NamedAxis}}) the third axis}}}
\section{Extends}{\code{\linkS4class{AxisSet2D}}}
\section{Methods}{\describe{\item{\code{\link[=print,AxisSet3D-method]{print}}}{\code{signature( = "AxisSet3D")}: #'}\item{\code{\link[=show,AxisSet3D-method]{show}}}{}}}

