\name{AxisSet4D-class}
\alias{AxisSet4D-class}
\alias{AxisSet4D}
\title{A four-dimensional axis set...}
\description{A four-dimensional axis set}
\section{Slots}{\describe{\item{\code{l}:}{(\code{\link{NamedAxis}}) the fourth axis}}}
\section{Extends}{\code{\linkS4class{AxisSet3D}}}
\section{Methods}{\describe{\item{\code{\link[=show,AxisSet4D-method]{show}}}{}}}

