\name{axisToIndex}
\alias{axisToIndex}
\title{Generic function to convert 1-dimensional real axis coordinates along a single axis dimension to an 1D index along the same axis}
\arguments{
  \item{x}{the object}

  \item{real}{the axis coordinates}

  \item{dimNum}{the dimension number of the axis (e.g.  1,
  2, 3)}
}
\value{
  a vector of axis indices
}
\description{
  Generic function to convert 1-dimensional real axis
  coordinates along a single axis dimension to an 1D index
  along the same axis

  axisToIndex
}

