\name{bounds-methods}
\alias{bounds-methods}
\alias{bounds}
\title{Generic function to extract the spatial bounds (origin + dim * spacing) of an image...}
\description{Generic function to extract the spatial bounds (origin + dim * spacing) of an image
param x the object}
\alias{bounds,BrainSpace-method}
\alias{bounds}
\alias{bounds,BrainData-method}
\alias{bounds}

