\name{BrainBucket-class}
\alias{BrainBucket-class}
\alias{BrainBucket}
\title{BrainBucket...}
\description{BrainBucket}
\details{a four-dimensional image this conists of a sequence of labelled image volumes backed by a list}
\section{Slots}{\describe{\item{\code{source}:}{}\item{\code{labels}:}{(\code{\link{character}}) the names of the sub-volumes contained in the bucket}\item{\code{data}:}{}}}
\section{Extends}{\code{\linkS4class{BrainVector}}}
\section{Methods}{\describe{\item{\code{\link[=names,BrainBucket-method]{names}}}{}\item{\code{\link[=[[,BrainBucket,character,missing-method]{[[}}}{}\item{\code{\link[=[[,BrainBucket,numeric,missing-method]{[[}}}{}\item{\code{\link[=eachVolume,BrainBucket,function,missing-method]{eachVolume}}}{}\item{\code{\link[=eachVolume,BrainBucket,function,logical-method]{eachVolume}}}{}}}

