% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{BrainData-class}
\alias{BrainData-class}
\title{BrainData}
\description{
Base class for brain image data
}
\section{Slots}{

\describe{
\item{\code{source}}{an instance of class \code{\linkS4class{BaseSource}} to store the source of the data}

\item{\code{space}}{an instance of class \code{\linkS4class{BrainSpace}} to represent the geometry of the data space}
}}

