\name{BrainData-class}
\alias{BrainData-class}
\alias{BrainData}
\title{Base class for brain image data...}
\description{Base class for brain image data}
\section{Slots}{\describe{\item{\code{source}:}{(\code{\link{BaseSource}}) an instance of class \code{\linkS4class{BaseSource}} to store the source of the data}\item{\code{space}:}{(\code{\link{BrainSpace}}) an instance of class \code{\linkS4class{BrainSpace}} to represent the geometry of the data space}}}
\section{Extends}{\code{\linkS4class{VIRTUAL}}}
\section{Methods}{\describe{\item{\code{\link[=ndim,BrainData-method]{ndim}}}{}\item{\code{\link[=dim,BrainData-method]{dim}}}{}\item{\code{\link[=space,BrainData-method]{space}}}{}\item{\code{\link[=spacing,BrainData-method]{spacing}}}{}\item{\code{\link[=as.matrix,BrainData-method]{as.matrix}}}{}\item{\code{\link[=as.array,BrainData-method]{as.array}}}{}\item{\code{\link[=as.vector,BrainData-method]{as.vector}}}{}\item{\code{\link[=bounds,BrainData-method]{bounds}}}{}\item{\code{\link[=axes,BrainData-method]{axes}}}{}\item{\code{\link[=origin,BrainData-method]{origin}}}{}\item{\code{\link[=trans,BrainData-method]{trans}}}{}\item{\code{\link[=inverseTrans,BrainData-method]{inverseTrans}}}{}}}

