\docType{class}
\name{BrainFileSource-class}
\alias{BrainFileSource-class}
\title{Base class for representing a data source for images. The purpose of this class is to provide a layer in between
low level IO and image loading functionality.}
\description{
Base class for representing a data source for images. The
purpose of this class is to provide a layer in between low
level IO and image loading functionality.
}
\section{Slots}{

\describe{
\item{\code{metaInfo}}{meta information for the data source}
}}

