\name{BrainFileSource-class}
\alias{BrainFileSource-class}
\alias{BrainFileSource}
\title{Base class for representing a data source for images.}
\description{Base class for representing a data source for images. The purpose of this class is to provide a layer in between 
low level IO and image loading functionality.}
\section{Slots}{\describe{\item{\code{metaInfo}:}{(\code{\link{FileMetaInfo}}) meta information for the data source}}}
\section{Extends}{\code{\linkS4class{BrainSource}}}

