\name{BrainMetaInfo-class}
\alias{BrainMetaInfo-class}
\alias{BrainMetaInfo}
\title{This class contains meta information from an image...}
\description{This class contains meta information from an image}
\section{Slots}{\describe{\item{\code{dataType}:}{(\code{\link{character}}) the data type code, e.g. FLOAT}\item{\code{Dim}:}{(\code{\link{numeric}}) image dimensions}\item{\code{spatialAxes}:}{(\code{\link{AxisSet3D}}) image axes for spatial dimensions (x,y,z)}\item{\code{additionalAxes}:}{(\code{\link{AxisSet}}) axes for dimensions > 3 (e.g. time, color band, direction)}\item{\code{spacing}:}{(\code{\link{numeric}}) voxel dimensions}\item{\code{origin}:}{(\code{\link{numeric}}) coordinate origin}\item{\code{label}:}{(\code{\link{character}}) name(s) of images}}}
\section{Extends}{\code{\linkS4class{BaseMetaInfo}}}

