\name{BrainMetaInfo-class}
\alias{BrainMetaInfo-class}
\alias{BrainMetaInfo}
\title{This class contains meta information from an image data file...}
\description{This class contains meta information from an image data file}
\section{Slots}{\describe{\item{\code{headerFile}:}{(\code{\link{character}}) name of the file containing meta information}\item{\code{dataFile}:}{(\code{\link{character}}) name of the file containing data}\item{\code{fileDescriptor}:}{(\code{\link{BrainFileDescriptor}}) descriptor of image file format}\item{\code{endian}:}{(\code{\link{character}}) byte order of data ('little' or 'big')}\item{\code{dataOffset}:}{(\code{\link{numeric}}) the number of bytes preceding the start of image data in data file}\item{\code{dataType}:}{(\code{\link{character}}) can be one of BYTE, SHORT, INT, FLOAT, DOUBLE -- multiple values allowed (must equal number of sub-images)}\item{\code{bytesPerElement}:}{(\code{\link{integer}}) number of bytes per element}\item{\code{Dim}:}{(\code{\link{numeric}}) image dimensions}\item{\code{spatialAxes}:}{(\code{\link{AxisSet3D}}) image axes for spatial dimensions (x,y,z)}\item{\code{additionalAxes}:}{(\code{\link{AxisSet}}) axes for dimensions > 3 (e.g. time, color band, direction)}\item{\code{spacing}:}{(\code{\link{numeric}}) voxel dimensions}\item{\code{origin}:}{(\code{\link{numeric}}) coordinate origin}\item{\code{label}:}{(\code{\link{character}}) name(s) of images contained in file}\item{\code{intercept}:}{(\code{\link{numeric}}) constant value added to image -- multiple values allowed (must equal numer of sub-images)}\item{\code{slope}:}{(\code{\link{numeric}}) image multiplier -- multiple values allowed (must equal numer of sub-images)}\item{\code{header}:}{(\code{\link{list}}) a list of format specific attributes}}}
\section{Extends}{\code{\linkS4class{BaseMetaInfo}}}
\section{Methods}{\describe{\item{\code{\link[=show,BrainMetaInfo-method]{show}}}{}}}

