\name{BrainSlice}
\alias{BrainSlice}
\title{BrainVolume}
\usage{
  BrainSlice(data, space, indices = NULL)
}
\arguments{
  \item{data}{data vector or matrix}

  \item{space}{an instance of class BrainSpace}

  \item{indices}{linear indices corresponding to data
  elements}
}
\description{
  Two-dimensional brain image

  BrainSlice constructor
}

