\name{BrainSlice}
\alias{BrainSlice}
\title{ creates a new "BrainSlice" object }
\description{
  function to create a new object of class "BrainSlice"
}
\usage{
BrainSlice(data, space, indices = NULL)
}

\arguments{
  \item{data}{ image values for the slice }
  \item{space}{ object of class "BrainSpace" describing the geometry of
    the slice }
  \item{indices}{ the 1d indices that the supplied data values
    correspond to }
}

\value{
  an object of class "BrainSlice"
  
}
\examples{
}
\keyword{manip}

