% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{BrainSource-class}
\alias{BrainSource-class}
\title{BrainSource}
\description{
Base class for representing a data source for images. The purpose of this class is to provide a layer in between
low level IO and image loading functionality.
}
\section{Slots}{

\describe{
\item{\code{metaInfo}}{meta information for the data source}
}}

