\name{BrainSpace-class}
\alias{BrainSpace-class}
\alias{BrainSpace}
\title{BrainSpace-class}
\section{Prototype}{\describe{\item{\emph{\code{Dim}} =}{111}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{origin}} =}{000}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{spacing}} =}{000}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{orientation}} =}{LPI}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{trans}} =}{1000010000100001}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{invTrans}} =}{1000010000100001}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{}} =}{}\item{\emph{\code{reptime}} =}{1}}}

