\name{BrainSpace-class}
\alias{BrainSpace}
\alias{BrainSpace-class}
\title{BrainSpace}
\usage{
  BrainSpace(Dim, origin = NULL, spacing = NULL,
    axes = NULL, trans = NULL)
}
\arguments{
  \item{Dim}{the dimensions}

  \item{origin}{the coordinate origin of the image space}

  \item{spacing}{the voxel dimensions}

  \item{axes}{the images axes specification}

  \item{trans}{the coordinate transformation associated
  with the image space}
}
\description{
  This class represents the geometry of a brain image

  Constructor function for BrainSpace class
}

