\name{BrainSpace}
\alias{BrainSpace}
\title{Constructor function for BrainSpace class}
\usage{
  BrainSpace(Dim, origin = NULL, spacing = NULL,
    axes = NULL, trans = NULL)
}
\arguments{
  \item{Dim}{the dimensions}

  \item{origin}{the coordinate origin of the image space}

  \item{spacing}{the voxel dimensions}

  \item{axes}{the images axes specification}

  \item{trans}{the coordinate transformation associated
  with the image space}
}
\description{
  Constructor function for BrainSpace class
}

