% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{BrainSurface-class}
\alias{BrainSurface-class}
\title{BrainSurface}
\description{
a three-dimensional surface consisting of a set of triangle vertices with one value per vertex.
}
\section{Slots}{

\describe{
\item{\code{source}}{the data source for the surface}

\item{\code{geometry}}{the surface geometry, an instance of \code{SurfaceGeometry}}

\item{\code{data}}{the 1-D vector of data value at each vertex of the mesh}
}}

