% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{BrainSurface-class}
\alias{BrainSurface-class}
\title{BrainSurface}
\description{
a three-dimensional surface consisting of a set of triangle vertices with one value per vertex.
}
\section{Slots}{

\describe{
\item{\code{source}}{the data source for the surface}

\item{\code{mesh}}{the underlying \code{tmesh3d} object}

\item{\code{vals}}{the value at each vertex of the mesh}
}}

