% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R, R/BrainSurface.R
\docType{class}
\name{BrainSurfaceSource-class}
\alias{BrainSurfaceSource}
\alias{BrainSurfaceSource-class}
\title{BrainSurfaceSource}
\usage{
BrainSurfaceSource(surfaceName, surfaceDataName, index = 1)
}
\arguments{
\item{surfaceName}{the name of the file containing the surface geometry.}

\item{surfaceDataName}{the name of the file containing the data values to be mapped to the surface.}
}
\description{
A class that is used to produce a \code{\linkS4class{BrainSurface}} instance

Constructor for BrainSurfaceSource
}
\section{Slots}{

\describe{
\item{\code{metaInfo}}{a \code{\linkS4class{SurfaceGeometryMetaInfo}} instance}

\item{\code{dataMetaInfo}}{a \code{\linkS4class{SurfaceDataMetaInfo}} instance}

\item{\code{index}}{the index offset into the surface data matrix}
}}

