% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R, R/BrainSurface.R
\docType{class}
\name{BrainSurfaceSource-class}
\alias{BrainSurfaceSource}
\alias{BrainSurfaceSource-class}
\title{BrainSurfaceSource}
\usage{
BrainSurfaceSource(surfaceGeom, surfaceDataName, indices = NULL)
}
\arguments{
\item{surfaceGeom}{the name of the file containing the surface geometry or a \code{SurfaceGeometry} instance}

\item{surfaceDataName}{the name of the file containing the data values to be mapped to the surface.}

\item{indices}{the indices to load from surface data matrix (if provided)}
}
\description{
A class that is used to produce a \code{\linkS4class{BrainSurface}} instance

Constructor for BrainSurfaceSource
}
\section{Slots}{

\describe{
\item{\code{geometry}}{a \code{\linkS4class{SurfaceGeometry}} instance}

\item{\code{dataMetaInfo}}{a \code{\linkS4class{SurfaceDataMetaInfo}} instance}

\item{\code{index}}{the column index of the surface map to be loaded.}
}}

