% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{BrainSurfaceVector-class}
\alias{BrainSurfaceVector-class}
\title{BrainSurfaceVector}
\description{
a three-dimensional surface consisting of a set of triangle vertices with multiple values per vertex.
}
\section{Slots}{

\describe{
\item{\code{source}}{the data source for the surface}

\item{\code{mesh}}{the underlying \code{mesh3d} object}

\item{\code{mat}}{a matrix of values where each column contains a vector of values over the surface nodes.}
}}

