% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{BrainSurfaceVectorSource-class}
\alias{BrainSurfaceVectorSource-class}
\title{BrainSurfaceVectorSource}
\description{
A class that is used to produce a \code{\linkS4class{BrainSurfaceVector}} instance
}
\section{Slots}{

\describe{
\item{\code{geometry}}{a \code{\linkS4class{SurfaceGeometry}} instance}

\item{\code{dataMetaInfo}}{a \code{\linkS4class{SurfaceDataMetaInfo}} instance}

\item{\code{indices}}{the column indices vector of the surface maps to be loaded}
}}

