\name{BrainVector-class}
\alias{BrainVector-class}
\alias{BrainVector}
\title{Four-dimensional brain image...}
\description{Four-dimensional brain image}
\section{Extends}{\code{\linkS4class{BrainData}}}
\section{Methods}{\describe{\item{\code{\link[=length,BrainVector-method]{length}}}{\code{signature( = “BrainVector”)}: length of 4th dimension}\item{\code{\link[=show,BrainVector-method]{show}}}{}\item{\code{\link[=eachVolume,BrainVector,function,missing-method]{eachVolume}}}{\code{signature(x = “BrainVector”, FUN = “function”, withIndex = “missing”)}: apply function to each volume in a BrainVector object}\item{\code{\link[=eachVolume,BrainVector,function,logical-method]{eachVolume}}}{\code{signature(x = “BrainVector”, FUN = “function”, withIndex = “logical”)}: apply function to each volume in a BrainVector object}\item{\code{\link[=takeVolume,BrainVector,numeric-method]{takeVolume}}}{\code{signature(x = “BrainVector”, i = “numeric”)}: takeVolume}\item{\code{\link[=eachSeries,BrainVector,function,missing-method]{eachSeries}}}{}\item{\code{\link[=concat,BrainVector,BrainVolume-method]{concat}}}{}\item{\code{\link[=concat,BrainVector,BrainVector-method]{concat}}}{}\item{\code{\link[=concat,BrainVector,BrainVector-method]{concat}}}{}\item{\code{\link[=seriesIter,BrainVector-method]{seriesIter}}}{}\item{\code{\link[=writeVector,BrainVector,character-method]{writeVector}}}{}\item{\code{\link[=indexToGrid,BrainVector,index-method]{indexToGrid}}}{}\item{\code{\link[=Arith,BrainVector,BrainVector-method]{Arith}}}{}\item{\code{\link[=Arith,BrainVector,BrainVector-method]{Arith}}}{}\item{\code{\link[=Arith,BrainVector,BrainVolume-method]{Arith}}}{}}}

