\name{BrainVectorSource-class}
\alias{BrainVectorSource-class}
\alias{BrainVectorSource}
\title{A class that is used to produce a \linkS4class{BrainVector} instance...}
\description{A class that is used to produce a \code{\linkS4class{BrainVector}} instance}
\section{Slots}{\describe{\item{\code{indices}:}{(\code{\link{integer}}) the index vector of the volumes to be loaded}}}
\section{Extends}{\code{\linkS4class{BrainSource}}}
\section{Methods}{\describe{\item{\code{\link[=loadData,BrainVectorSource-method]{loadData}}}{\code{signature( = “BrainVectorSource”)}: Load data from a \code{\linkS4class{BrainVectorSource}}}\item{\code{\link[=show,BrainVectorSource-method]{show}}}{}}}

