\name{BrainVectorSource}
\alias{BrainVectorSource}
\title{Construct a \linkS4class{BrainVectorSource} object...}
\usage{BrainVectorSource(input, indices)
}
\description{Construct a \code{\linkS4class{BrainVectorSource}} object}
\arguments{\item{indices}{the subset of volume indices to load -- if \code{NULL} then all volumes will be loaded}
}

