\name{BrainVolume}
\alias{BrainVolume}
\title{Construct a \linkS4class{BrainVolume} instance, using default (dense) implementation...}
\usage{BrainVolume(data, space, source)
}
\description{Construct a \code{\linkS4class{BrainVolume}} instance, using default (dense) implementation}
\arguments{\item{data}{a three-dimensional \code{array}}
\item{space}{an instance of class \code{\linkS4class{BrainSpace}}}
\item{source}{an instance of class \code{\linkS4class{BrainSource}}}
}

