% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R, R/BrainVolume.R
\docType{class}
\name{BrainVolumeSource-class}
\alias{BrainVolumeSource-class}
\alias{BrainVolumeSource}
\title{BrainVolumeSource
		
A class is used to produce a \code{\linkS4class{BrainVolume}} instance}
\usage{
BrainVolumeSource(input, index = 1)
}
\arguments{
\item{input}{the input file name}

\item{index}{the image subvolume index}
}
\description{
BrainVolumeSource
		
A class is used to produce a \code{\linkS4class{BrainVolume}} instance

Constructor for BrainVolumeSource
}
\section{Slots}{

\describe{
\item{\code{index}}{the index of the volume to be read -- must be of length 1.}
}}

