\name{BrainVolumeSource-class}
\alias{BrainVolumeSource}
\alias{BrainVolumeSource-class}
\title{A class is used to produce a \code{\linkS4class{BrainVolume}} instance}
\usage{
  BrainVolumeSource(input, index = 1)
}
\arguments{
  \item{input}{the input file name}

  \item{index}{the image subvolume index}
}
\description{
  A class is used to produce a
  \code{\linkS4class{BrainVolume}} instance

  Constructor for BrainVolumeSource
}

