% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryIO.R
\docType{methods}
\name{close,BinaryReader-method}
\alias{close,BinaryReader-method}
\alias{close,BinaryWriter-method}
\title{close}
\usage{
\S4method{close}{BinaryReader}(con)

\S4method{close}{BinaryWriter}(con)
}
\arguments{
\item{con}{the object to close}
}
\description{
close
}
