% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{close,BinaryReader-method}
\alias{close,BinaryReader-method}
\alias{close,BinaryWriter-method}
\title{close}
\usage{
\S4method{close}{BinaryReader}(con)

\S4method{close}{BinaryWriter}(con)
}
\arguments{
\item{con}{a connection.}
}
\description{
close

close
}

