\docType{methods}
\name{clusterCenters}
\alias{clusterCenters}
\alias{clusterCenters,ClusteredBrainVolume,matrix,missing-method}
\title{clusterCenters}
\usage{
clusterCenters(x, features, FUN)

\S4method{clusterCenters}{ClusteredBrainVolume,matrix,missing}(x, features)
}
\arguments{
  \item{x}{the object to extract cluster centers from}
}
\description{
clusterCenters

extract cluster centers in a ClusteredBrainVolume
}

