\name{concat-methods}
\alias{concat-methods}
\alias{concat}
\title{Concatenate two objects...}
\description{Concatenate two objects}
\arguments{\item{x}{the first object}
\item{y}{the second object}
\item{...}{additonal objects}
}
\alias{concat,BrainVector,BrainVolume-method}
\alias{concat}
\alias{concat,BrainVector,BrainVector-method}
\alias{concat}
\alias{concat,DenseBrainVolume,DenseBrainVolume-method}
\alias{concat}
\note{dimensions of x and y must be equal}
\alias{concat,SparseBrainVector,SparseBrainVector-method}
\alias{concat}

