\name{concat-methods}
\alias{concat,BrainVector,BrainVolume-method}
\alias{concat}
\title{concat,BrainVector,BrainVolume-method}
\alias{concat,BrainVector,BrainVector-method}
\alias{concat}
\alias{concat,DenseBrainVolume,DenseBrainVolume-method}
\alias{concat}
\description{concatenate two BrainVolumes}
\note{dimensions of x and y must be equal}
\arguments{\item{x}{the first volume}
\item{y}{the second volume}
\item{...}{extra arguments of class BrainVolume or other concatenable objects}
}
\alias{concat,SparseBrainVector,SparseBrainVector-method}
\alias{concat}

