\name{concat}
\alias{concat}
\title{Concatenate two objects}
\arguments{
  \item{x}{the first object}

  \item{y}{the second object}

  \item{...}{additonal objects}

  \item{x}{the first volume}

  \item{y}{the second volume}

  \item{...}{extra arguments of class BrainVolume or other
  concatenable objects}
}
\description{
  Concatenate two objects

  concat

  concat

  concat

  concatenate two BrainVolumes
}
\note{
  dimensions of x and y must be equal
}

