\docType{methods}
\name{connComp}
\alias{connComp}
\alias{connComp,BrainVolume,BrainVolume-method}
\title{Find connected components}
\arguments{
  \item{x}{the image object}

  \item{...}{additonal arguments}
}
\description{
  Find connected components

  find connected components in BrainVolume
}

