% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conncomp.R
\name{connComp3D}
\alias{connComp3D}
\title{Extract connected components from a 3D mask}
\usage{
connComp3D(mask)
}
\arguments{
\item{mask}{a 3D binary array}
}
\value{
a two-element list of the connected components (cluster 'index' and cluster 'size')
The first element 'index' is a 3D array containing the cluster index of the connected component for each voxel.
The second element 'size' is a 3D array consisting of the size of the connected component inhabited by each voxel.
}
\description{
Extract connected components from a 3D mask
}

