% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conncomp.R
\name{connComp3D}
\alias{connComp3D}
\title{Extract connected components from a 3D mask}
\usage{
connComp3D(mask)
}
\arguments{
\item{mask}{a 3D binary array}
}
\value{
a two-element list of the connected components (cluster index and cluster sizes)
}
\description{
Extract connected components from a 3D mask
}

