\name{connComp3D}
\alias{connComp3D}
\title{Extract connected components from a 3D mask...}
\usage{connComp3D(mask)
}
\description{Extract connected components from a 3D mask}
\value{a two-element list of the connected components (cluster index and cluster sizes)}
\arguments{\item{mask}{a 3D binary array}
}

