\name{coords-methods}
\alias{coords-methods}
\alias{coords}
\title{Extract coordinates...}
\description{Extract coordinates}
\arguments{\item{x}{the object to extract coordinates from}
}
\alias{coords,BrainRegion3D-method}
\alias{coords}
\usage{\S4method{coords}{BrainRegion3D}(x)
}
\alias{coords,-method}
\alias{coords}
\alias{coords,SparseBrainVector-method}
\alias{coords}

