\name{coordToGrid}
\alias{coordToGrid}
\title{Generic function to convert N-dimensional real world coordinates to grid coordinates}
\arguments{
  \item{x}{the object}

  \item{coords}{a matrix of real world coordinates}
}
\value{
  a matrix of grid coordinates
}
\description{
  Generic function to convert N-dimensional real world
  coordinates to grid coordinates

  coordToGrid

  coordToGrid
}

