\name{coordToIndex}
\alias{coordToIndex}
\title{Generic function to convert N-dimensional real world coordinates to 1D indices}
\arguments{
  \item{x}{the object}

  \item{coords}{a matrix of real world coordinates}
}
\value{
  a vector of indices
}
\description{
  Generic function to convert N-dimensional real world
  coordinates to 1D indices

  coordToIndex

  coordToIndex
}

