\name{dataFile}
\alias{dataFile}
\title{Generic function to get the name of the data file, given a file name and a \code{\linkS4class{BrainFileDescriptor}} instance.}
\arguments{
  \item{x}{descriptor instance}

  \item{fileName}{file name to be stripped of its
  extension}
}
\value{
  the correct header name
}
\description{
  Generic function to get the name of the data file, given
  a file name and a
  \code{\linkS4class{BrainFileDescriptor}} instance.
}

