\name{dataFile-methods}
\alias{dataFile-methods}
\alias{dataFile}
\title{Generic function to get the name of the data file, given a file name and a \linkS4class{BrainFileDescriptor} instance.}
\description{Generic function to get the name of the data file, given a file name and a \code{\linkS4class{BrainFileDescriptor}} instance.}
\value{\code{dataFile-methods}: the correct header name

}
\arguments{\item{x}{descriptor instance}
\item{fileName}{file name to be stripped of its extension}
}
\alias{dataFile,BrainFileDescriptor,character-method}
\alias{dataFile}

