\name{dataFileMatches-methods}
\alias{dataFileMatches-methods}
\alias{dataFileMatches}
\title{Generic function to test whether a file name conforms to the given a \linkS4class{BrainFileDescriptor} instance.}
\description{Generic function to test whether a file name conforms to the given a \code{\linkS4class{BrainFileDescriptor}} instance.
Will test for match to data file only}
\value{\code{dataFileMatches-methods}: TRUE for match, FALSE otherwise

}
\arguments{\item{x}{object for which the file name is to matched to}
\item{fileName}{file name to be matched}
}
\alias{dataFileMatches,BrainFileDescriptor,character-method}
\alias{dataFileMatches}
\alias{dataFile,BrainFileDescriptor,character-method}
\alias{dataFile}

