\name{dataReader}
\alias{dataReader}
\title{Generic function to create data reader}
\arguments{
  \item{x}{an object specifying the infromation required to
  produce the reader}

  \item{offset}{the byte offset (number of bytes to skip
  before reading)}
}
\description{
  Generic function to create data reader
}

