\name{dataReader-methods}
\alias{dataReader-methods}
\alias{dataReader}
\title{Generic function to create data reader...}
\description{Generic function to create data reader}
\arguments{\item{x}{an object specifying the infromation required to produce the reader}
\item{offset}{the byte offset (number of bytes to skip before reading)}
}

