\name{DenseBrainVector-class}
\alias{DenseBrainVector-class}
\alias{DenseBrainVector}
\title{Four-dimensional brain image, backed by an array...}
\description{Four-dimensional brain image, backed by an array}
\section{Extends}{\code{\linkS4class{BrainVector}}, \code{\linkS4class{array}}}
\alias{DenseBrainVector}
\usage{DenseBrainVector(data, space, source, label="")
}
\details{\code{DenseBrainVector}: constructor function for class \code{\linkS4class{DenseBrainVector}}

}
\value{\code{DenseBrainVector}: \code{\linkS4class{DenseBrainVector}} instance

}
\arguments{\item{data}{a 4-dimensonal \code{array}}
\item{space}{a \code{\linkS4class{BrainSpace}} object}
\item{source}{an optional \code{\linkS4class{BrainSource}} object}
\item{label}{a label of type \code{character}}
}

