\name{DenseBrainVolume-class}
\alias{DenseBrainVolume-class}
\alias{DenseBrainVolume}
\title{Three-dimensional brain image, backed by an array...}
\description{Three-dimensional brain image, backed by an \code{array}}
\section{Extends}{\code{\linkS4class{BrainVolume}}, \code{\linkS4class{array}}}
\section{Methods}{\describe{\item{\code{\link[=concat,DenseBrainVolume,DenseBrainVolume-method]{concat}}}{\code{signature(x = “DenseBrainVolume”, y = “DenseBrainVolume”)}: concatenate two BrainVolumes}\item{\code{\link[=concat,DenseBrainVolume,DenseBrainVolume-method]{concat}}}{\code{signature(x = “DenseBrainVolume”, y = “DenseBrainVolume”)}: concatenate two BrainVolumes}}}

