% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{dim,BrainData-method}
\alias{dim,BrainData-method}
\alias{dim,BrainSpace,ANY-method}
\alias{dim,BrainSpace-method}
\alias{dim,FileMetaInfo,ANY-method}
\alias{dim,FileMetaInfo-method}
\title{dim}
\usage{
\S4method{dim}{BrainData}(x)

\S4method{dim}{FileMetaInfo}(x)

\S4method{dim}{BrainSpace}(x)
}
\arguments{
\item{x}{an \R object, for example a matrix, array or data frame.}
}
\description{
dim

dim
}

