\docType{methods}
\name{dropDim}
\alias{dropDim}
\alias{dropDim,AxisSet2D,missing-method}
\alias{dropDim,AxisSet2D,numeric-method}
\alias{dropDim,AxisSet3D,missing-method}
\alias{dropDim,AxisSet3D,numeric-method}
\alias{dropDim,BrainSpace,missing-method}
\title{Generic function to drop a dimension from an object}
\usage{
dropDim(x, dimnum)

\S4method{dropDim}{AxisSet2D,numeric}(x, dimnum)

\S4method{dropDim}{AxisSet2D,missing}(x)

\S4method{dropDim}{AxisSet3D,numeric}(x, dimnum)

\S4method{dropDim}{AxisSet3D,missing}(x)

\S4method{dropDim}{BrainSpace,missing}(x)
}
\arguments{
  \item{x}{a dimensioned object}
}
\description{
Generic function to drop a dimension from an object

dropDim

dropDim

dropDim

dropDim

dropDim
}

