\name{eachSlice-methods}
\alias{eachSlice-methods}
\alias{eachSlice}
\title{Generic functions to apply a function to each (2D) slice of an image...}
\description{Generic functions to apply a function to each (2D) slice of an image}
\arguments{\item{x}{the object}
\item{FUN}{a \code{function} taking one or two arguments (depending on the value of \code{withIndex}}
\item{withIndex}{whether the index of the slice is supplied as the second argument to the function}
}
\alias{eachSlice,BrainVolume,function,missing-method}
\alias{eachSlice}
\alias{eachSlice,BrainVolume,function,logical-method}
\alias{eachSlice}

