\name{eachSlice}
\alias{eachSlice}
\title{Generic functions to apply a function to each (2D) slice of an image}
\arguments{
  \item{x}{the object}

  \item{FUN}{a \code{function} taking one or two arguments
  (depending on the value of \code{withIndex}}

  \item{withIndex}{whether the index of the slice is
  supplied as the second argument to the function}
}
\description{
  Generic functions to apply a function to each (2D) slice
  of an image

  eachSlice

  eachSlice
}

