\name{eachVolume-methods}
\alias{eachVolume-methods}
\alias{eachVolume}
\title{Generic function to apply a function to each volume of a four-dimensional image...}
\description{Generic function to apply a function to each volume of a four-dimensional image}
\arguments{\item{x}{four-dimensional image}
\item{FUN}{a \code{function} taking one or two arguments (depending on the value of \code{withIndex}}
\item{withIndex}{whether the index of the volume supplied as the second argument to the function}
}
\alias{eachVolume,BrainVector,function,missing-method}
\alias{eachVolume}
\alias{eachVolume,BrainBucket,function,missing-method}
\alias{eachVolume}
\alias{eachVolume,BrainBucket,function,logical-method}
\alias{eachVolume}
\alias{eachVolume,BrainVector,function,logical-method}
\alias{eachVolume}

