\name{eachVolume}
\alias{eachVolume}
\title{Generic function to apply a function to each volume of a four-dimensional image}
\arguments{
  \item{x}{four-dimensional image}

  \item{FUN}{a \code{function} taking one or two arguments
  (depending on the value of \code{withIndex}}

  \item{withIndex}{whether the index of the volume supplied
  as the second argument to the function}
}
\description{
  Generic function to apply a function to each volume of a
  four-dimensional image

  eachVolume

  eachVolume

  eachVolume

  eachVolume

  eachVolume
}

