\name{fileMatches-methods}
\alias{fileMatches-methods}
\alias{fileMatches}
\title{Generic function to test whether a file name conforms to the given \linkS4class{BrainFileDescriptor} instance.}
\description{Generic function to test whether a file name conforms to the given \code{\linkS4class{BrainFileDescriptor}} instance.
Will test for match to either header file or data file}
\value{\code{fileMatches-methods}: TRUE for match, FALSE otherwise

}
\arguments{\item{x}{object for which the file name is to matched to}
\item{fileName}{file name to be matched}
}
\alias{fileMatches,BrainFileDescriptor,character-method}
\alias{fileMatches}

