\name{fileMatches}
\alias{fileMatches}
\title{Generic function to test whether a file name conforms to the given \code{\linkS4class{BrainFileDescriptor}} instance.
Will test for match to either header file or data file}
\arguments{
  \item{x}{object for which the file name is to matched to}

  \item{fileName}{file name to be matched}
}
\value{
  TRUE for match, FALSE otherwise
}
\description{
  Generic function to test whether a file name conforms to
  the given \code{\linkS4class{BrainFileDescriptor}}
  instance. Will test for match to either header file or
  data file
}

