% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{FileMetaInfo-class}
\alias{FileMetaInfo-class}
\alias{NIfTIMetaInfo-class}
\alias{AFNIMetaInfo-class}
\title{FileMetaInfo}
\description{
This class contains meta information from an image data file

This class contains meta information for a NIfTI image file

This class contains meta information for a AFNI image file
}
\section{Slots}{

\describe{
\item{\code{headerFile}}{name of the file containing meta information}

\item{\code{dataFile}}{name of the file containing data}

\item{\code{fileDescriptor}}{descriptor of image file format}

\item{\code{endian}}{byte order of data ('little' or 'big')}

\item{\code{dataOffset}}{the number of bytes preceding the start of image data in data file}

\item{\code{bytesPerElement}}{number of bytes per element}

\item{\code{intercept}}{constant value added to image -- multiple values allowed (must equal numer of sub-images)}

\item{\code{slope}}{image multiplier -- multiple values allowed (must equal numer of sub-images)}

\item{\code{header}}{a list of format specific attributes}

\item{\code{nifti_header}}{a \code{list} of attributes specific to the NIfTI file format}

\item{\code{afni_header}}{a list of attributes specific to the AFNI file format}

\item{\code{afni_header}}{a \code{list} of attributes specific to the AFNI file format}
}}

