\name{fill}
\alias{fill}
\title{Generic function to map values from one set to another using a user-supplied lookup table}
\arguments{
  \item{x}{the object to map values from}

  \item{lookup}{the lookup table}
}
\value{
  a new object where the original values have been filled
  in with the values in the lookup table
}
\description{
  Generic function to map values from one set to another
  using a user-supplied lookup table
}

