\name{gridToIndex}
\alias{gridToIndex}
\title{Generic function to convert N-dimensional grid coordinate to 1D indices}
\arguments{
  \item{x}{the object}

  \item{coords}{a matrix where each row is a corodinate or
  a vector of length N}
}
\value{
  a vector of indices
}
\description{
  Generic function to convert N-dimensional grid coordinate
  to 1D indices

  gridToIndex

  gridToIndex
}

