\name{gridToIndex-methods}
\docType{methods}
\alias{gridToIndex-methods}
\alias{gridToIndex,BrainSlice,matrix-method}
\alias{gridToIndex,BrainVolume,matrix-method}
\title{ map grid coordinate to 1d index coordinate }
\description{
 Methods for function \code{gridToIndex}  
}
\section{Methods}{
\describe{

\item{x = "BrainSlice", coords = "matrix"}{ map 2-d coords to 1-d index }

\item{x = "BrainVolume", coords = "matrix"}{ map 3-d coordinates to 1-d index }
}}
\keyword{methods}

