\name{gridToIndex-methods}
\alias{gridToIndex-methods}
\alias{gridToIndex}
\title{Generic function to convert N-dimensional grid coordinate to 1D indices...}
\description{Generic function to convert N-dimensional grid coordinate to 1D indices}
\value{a vector of indices}
\arguments{\item{x}{the object}
\item{coords}{a matrix where each row is a corodinate or a vector of length N}
}

