\name{headerFile-methods}
\docType{methods}
\alias{headerFile-methods}
\alias{headerFile,BrainHeader-method}
\title{ Methods for the generic function headerFile }
\description{
  returns the name of the header file given a \code{"BrainFile"} instance.
}
\section{Methods}{
\describe{

\item{x = "BrainHeader"}{ returns the name of the header file }
}}
\keyword{methods}

