\name{headerFile-methods}
\alias{headerFile-methods}
\alias{headerFile}
\title{Generic function to get the name of the header file, given a file name and a \linkS4class{BrainFileDescriptor} instance.}
\description{Generic function to get the name of the header file, given a file name and a \code{\linkS4class{BrainFileDescriptor}} instance.}
\value{the correct header name}
\arguments{\item{x}{descriptor instance}
\item{fileName}{file name to be stripped of its extension}
}

