\name{headerFileMatches-methods}
\alias{headerFileMatches-methods}
\alias{headerFileMatches}
\title{Generic function to test whether a file name conforms to the given \linkS4class{BrainFileDescriptor} instance.}
\description{Generic function to test whether a file name conforms to the given \code{\linkS4class{BrainFileDescriptor}} instance.
Will test for match to header file only}
\value{\code{headerFileMatches-methods}: TRUE for match, FALSE otherwise

}
\arguments{\item{x}{object for which the file name is to matched to}
\item{fileName}{file name to be matched}
}
\alias{headerFileMatches,BrainFileDescriptor,character-method}
\alias{headerFileMatches}

