% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{image,BrainVolume-method}
\alias{image,BrainVolume-method}
\alias{image,Layer-method}
\alias{image,Overlay-method}
\title{image}
\usage{
\S4method{image}{BrainVolume}(x, slice, col = heat.colors(128, alpha = 1),
  zero.col = "#00000000")

\S4method{image}{Overlay}(x, zpos, axis = 3)

\S4method{image}{Layer}(x, zpos, axis = 3)
}
\arguments{
\item{slice}{the voxel index of the slice to display}

\item{col}{a color map}

\item{zero.col}{the color to use when the value is 0 (e.g background color)}

\item{zpos}{the z coordinate}

\item{axis}{the axis index}

\item{x}{locations of grid lines at which the values in \code{z} are
    measured.  These must be finite, non-missing and in (strictly)
    ascending order.  By default, equally
    spaced values from 0 to 1 are used.  If \code{x} is a \code{list},
    its components \code{x$x} and \code{x$y} are used for \code{x}
    and \code{y}, respectively. If the list has component \code{z} this
    is used for \code{z}.}
}
\description{
image

image

image
}

