% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Display.R
\docType{methods}
\name{image,BrainVolume-method}
\alias{image,BrainVolume-method}
\alias{image,Layer-method}
\alias{image,Overlay-method}
\title{image}
\usage{
\S4method{image}{BrainVolume}(x, slice, col = gray((0:255)/255, alpha = 1),
  zero.col = "#000000", axis = 3, ...)

\S4method{image}{Overlay}(x, zpos, axis = 3)

\S4method{image}{Layer}(x, zpos, axis = 3)
}
\arguments{
\item{slice}{the voxel index of the slice to display}

\item{col}{a color map}

\item{zero.col}{the color to use when the value is 0 (e.g background color)}

\item{axis}{the index of the z axis (1, 2, or 3)}

\item{zpos}{the z coordinate}

\item{axis}{the axis index}
}
\description{
image

image

image
}

