\docType{methods}
\name{image,BrainVolume-method}
\alias{image,BrainVolume-method}
\alias{image,Layer-method}
\alias{image,Overlay-method}
\title{image}
\usage{
\S4method{image}{BrainVolume}(x, slice, col = heat.colors(128, alpha = 1),
  zero.col = "#00000000")

\S4method{image}{Overlay}(x, zpos, axis = 3)

\S4method{image}{Layer}(x, zpos, axis = 3)
}
\arguments{
  \item{x}{locations of grid lines at which the values in
  \code{z} are measured.  These must be finite, non-missing
  and in (strictly) ascending order.  By default, equally
  spaced values from 0 to 1 are used.  If \code{x} is a
  \code{list}, its components \code{x$x} and \code{x$y} are
  used for \code{x} and \code{y}, respectively. If the list
  has component \code{z} this is used for \code{z}.}

  \item{col}{a list of colors such as that generated by
  \code{\link{rainbow}}, \code{\link{heat.colors}},
  \code{\link{topo.colors}}, \code{\link{terrain.colors}}
  or similar functions.}
}
\description{
image

image

image
}

