% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{imageGrid}
\alias{imageGrid}
\title{imageGrid}
\usage{
imageGrid(layer, gridDim = c(3, 3), zstart, zend, panelSize = 3,
  panelUnit = "inches", interpolate = FALSE, fontCol = "red")
}
\arguments{
\item{layer}{the layer to display}

\item{gridDim}{the dimensions of the 2D grid montage}

\item{zstart}{the z coordinate of the first slice}

\item{zend}{the z coordinate of the last slice}

\item{panelSize}{the size of each panel in the montage (default unit is inches)}

\item{panelUnit}{the unit for the panel size (default is inches)}

\item{interpolate}{whether to interpolate pixel values}

\item{fontCol}{color of labels indicating slice level}
}
\description{
Display a set of mages slices in a 2D montage
}

