\name{IndexLookupVolume-class}
\alias{IndexLookupVolume-class}
\alias{IndexLookupVolume}
\title{Three-dimensional brain image that can be used as a map between 1D grid indices and a table of values...}
\description{Three-dimensional brain image that can be used as a map between 1D grid indices and a table of values
Currently used in the \code{\linkS4class{SparseBrainVector}} class.}
\section{Extends}{\code{\linkS4class{BrainVolume}}}
\section{Methods}{\describe{\item{\code{\link[=initialize,IndexLookupVolume-method]{initialize}}}{}\item{\code{\link[=indices,IndexLookupVolume-method]{indices}}}{}\item{\code{\link[=lookup,IndexLookupVolume,numeric-method]{lookup}}}{}\item{\code{\link[=space,IndexLookupVolume-method]{space}}}{}}}

