\name{indexToCoord}
\alias{indexToCoord}
\title{Generic function to convert 1D indices to N-dimensional real world coordinates}
\arguments{
  \item{x}{the object}

  \item{idx}{the 1D indices}
}
\value{
  a matrix of real coordinates
}
\description{
  Generic function to convert 1D indices to N-dimensional
  real world coordinates

  indexToCoord

  indexToCoord
}

