\name{indexToGrid}
\alias{indexToGrid}
\title{Generic function to convert 1D indices to N-dimensional grid coordinates}
\arguments{
  \item{x}{the object}

  \item{idx}{the 1D indices}
}
\value{
  a matrix of grid coordinates
}
\description{
  Generic function to convert 1D indices to N-dimensional
  grid coordinates

  indexToGrid

  indexToGrid

  indexToGrid

  indexToGrid
}

