\name{indexToGrid-methods}
\alias{indexToGrid-methods}
\alias{indexToGrid}
\title{Generic function to convert a 1D index to N-dimensional grid coordinate...}
\description{Generic function to convert a 1D index to N-dimensional grid coordinate}
\value{a matrix of grid coordinates}
\arguments{\item{x}{the object}
\item{idx}{the 1D indices}
}

