\name{indexToGrid}
\alias{indexToGrid}
\title{Generic function to convert a 1D index to N-dimensional grid coordinate}
\arguments{
  \item{x}{the object}

  \item{idx}{the 1D indices}
}
\value{
  a matrix of grid coordinates
}
\description{
  Generic function to convert a 1D index to N-dimensional
  grid coordinate

  indexToGrid

  indexToGrid

  indexToGrid

  indexToGrid
}

