% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{indices}
\alias{indices}
\alias{indices,IndexLookupVolume-method}
\alias{indices,ROIVolume-method}
\alias{indices,SparseBrainVector-method}
\title{Extract indices}
\usage{
indices(x)

\S4method{indices}{ROIVolume}(x)

\S4method{indices}{IndexLookupVolume}(x)

\S4method{indices}{SparseBrainVector}(x)
}
\arguments{
\item{x}{the object to extract indices}
}
\description{
Extract indices

indices

indices

indices
}

