% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainRegion3D.R, R/SparseBrainVector.R, R/IndexLookupVolume.R
\docType{methods}
\name{indices}
\alias{indices}
\alias{indices,IndexLookupVolume-method}
\alias{indices,ROISurface-method}
\alias{indices,ROISurfaceVector-method}
\alias{indices,ROIVector-method}
\alias{indices,ROIVolume-method}
\alias{indices,SparseBrainVector-method}
\title{Extract indices}
\usage{
indices(x)

\S4method{indices}{ROIVolume}(x)

\S4method{indices}{ROIVector}(x)

\S4method{indices}{ROISurface}(x)

\S4method{indices}{ROISurfaceVector}(x)

\S4method{indices}{SparseBrainVector}(x)

\S4method{indices}{IndexLookupVolume}(x)
}
\arguments{
\item{x}{the object to extract indices}
}
\description{
Extract indices

indices
}

