\name{indices-methods}
\docType{methods}
\alias{indices-methods}
\alias{indices,IndexLookupVolume-method}
\alias{indices,SparseBrainVector-method}
\alias{indices,TiledBrainVector-method}
\title{ extract the indices associated with an object }
\description{
 Methods for function \code{indices}  
}
\section{Methods}{
\describe{

\item{x = "IndexLookupVolume"}{ extract the indices of an \code{IndexLookupVolume} }

\item{x = "SparseBrainVector"}{ extract the indices of an \code{SparseBrainVector}  }

\item{x = "TiledBrainVector"}{ extract the indices of an \code{TiledBrainVector} }
}}
\keyword{methods}

