\docType{methods}
\name{indices}
\alias{indices}
\alias{indices,ROIVolume,ANY-method}
\alias{indices,SparseBrainVector,ANY-method}
\title{Extract indices}
\arguments{
  \item{x}{the object to extract indices}

  \item{x}{an ROIVolume}
}
\description{
  Extract indices

  indices

  indices

  indices
}

