% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainSpace.R
\docType{methods}
\name{inverseTrans}
\alias{inverseTrans}
\alias{inverseTrans,BrainData-method}
\alias{inverseTrans,BrainSpace-method}
\title{Generic getter to extract inverse image coordinate transformation}
\usage{
inverseTrans(x)

\S4method{inverseTrans}{BrainSpace}(x)

\S4method{inverseTrans}{BrainData}(x)
}
\arguments{
\item{x}{an object}
}
\description{
Generic getter to extract inverse image coordinate transformation
}
\examples{
bspace <- BrainSpace(c(10,10,10), c(2,2,2))
itrans <- inverseTrans(bspace)
identical(trans(bspace) \%*\% inverseTrans(bspace), diag(4))
}

