\name{inverseTrans-methods}
\alias{inverseTrans-methods}
\alias{inverseTrans}
\title{Generic getter to extract inverse image coordinate transformation...}
\description{Generic getter to extract inverse image coordinate transformation}
\alias{inverseTrans,BrainSpace-method}
\alias{inverseTrans}
\alias{inverseTrans,BrainData-method}
\alias{inverseTrans}

