% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{class}
\name{Kernel-class}
\alias{Kernel-class}
\title{Kernel}
\description{
A class representing an image kernel
}
\section{Slots}{

\describe{
\item{\code{width}}{the width in voxels of the kernel}

\item{\code{weights}}{the kernel weights}

\item{\code{voxels}}{the relative voxel coordinates of the kernel}

\item{\code{coords}}{the relative real coordinates of the kernel}
}}

