% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{class}
\name{Layer-class}
\alias{Layer-class}
\title{Layer}
\description{
A class used for displaying 2D images with color maps
}
\section{Slots}{

\describe{
\item{\code{vol}}{the \code{BrainVolume} that provides the data for the layer.}

\item{\code{colorMap}}{a character vector of colors in hexadecimal rgb format.
Can be generated by calls to \code{rainbow}, \code{heat.colors}, \code{topo.colors}, \code{terrain.colors} or similar functions.}

\item{\code{thresh}}{cut-off value above which vlaues will be made transparent.}

\item{\code{axis}}{the axis index of perpendicular to the xy plane (option: 1,2,3; default is 3)}

\item{\code{zero.col}}{the color pixels with intensity of zero. This value overrides the color from the slot \code{colorMap}}
}}

