% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{Layer}
\alias{Layer}
\title{Layer}
\usage{
Layer(vol, colorMap = gray((0:255)/255, alpha = 1), thresh = c(0, 0),
  axis = 3, zero.col = "#000000")
}
\arguments{
\item{vol}{volume instance of \code{\linkS4class{BrainVolume}}}

\item{colorMap}{a lookup table defining mapping from image intensity values to colors.}

\item{thresh}{a range (min,max) defining the threshold window for determining image opacity.}

\item{axis}{the axis index of the axis perpendicular to the xy plane (options: 1,2,3; default is 3)}

\item{zero.col}{the color used when the value is zero.}
}
\value{
an object of class \code{Layer}
}
\description{
create a \code{\linkS4class{Layer}} object
}

