\name{loadBucket}
\alias{loadBucket}
\title{loadBucket...}
\usage{loadBucket(fileName, pattern, indices)
}
\description{loadBucket}
\details{load a BrainBucket object from file}
\arguments{\item{fileName}{the name of the file to load}
\item{pattern}{optional regular expression used to filter the sub-volumes using associated labels}
\item{indices}{optional set of sub-volume indices to load}
}

