\name{loadData-methods}
\alias{loadData-methods}
\alias{loadData}
\title{Generic function to load data from a data source...}
\description{Generic function to load data from a data source}
\arguments{\item{x}{a data source}
}
\alias{loadData,BrainVectorSource-method}
\alias{loadData}
\value{\code{loadData,BrainVectorSource-method}: an instance of class \code{\linkS4class{BrainVector}}

\code{loadData,BrainBucketSource-method}: an instance of class \code{\linkS4class{BrainVolume}}


\code{loadData,SparseBrainVectorSource-method}: an instance of class \code{\linkS4class{SparseBrainVector}}


}
\alias{loadData,BrainBucketSource-method}
\alias{loadData}
\docType{methods}
\alias{loadData,BrainVolumeSource-method}
\alias{loadData}
\alias{loadData,SparseBrainVectorSource-method}
\alias{loadData}

