\docType{methods}
\name{loadData}
\alias{loadData}
\title{Generic function to load data from a data source}
\arguments{
  \item{x}{a data source}

  \item{x}{an instance of class
  \code{\linkS4class{BrainVectorSource}}}

  \item{x}{an instance of class
  \code{\linkS4class{BrainBucketSource}}}
}
\value{
  an instance of class \code{\linkS4class{BrainVector}}

  an instance of class \code{\linkS4class{BrainVolume}}
}
\description{
  Generic function to load data from a data source

  loadData

  Load data from a \code{\linkS4class{BrainVectorSource}}

  Load data from a \code{\linkS4class{BrainBucketSource}}

  load a BrainVolume
}

