\name{loadData-methods}
\alias{loadData-methods}
\alias{loadData}
\title{Generic function to load data from a data source...}
\description{Generic function to load data from a data source}
\arguments{\item{x}{a data source}
}
\alias{loadData,SparseBrainVectorSource-method}
\alias{loadData}
\value{\code{loadData,SparseBrainVectorSource-method}: an instance of class \code{\linkS4class{SparseBrainVector}}

}

