% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\name{loadFSSurface}
\alias{loadFSSurface}
\title{load Freesurfer ascii surface}
\usage{
loadFSSurface(metaInfo)
}
\arguments{
\item{metaInfo}{instance of type \code{FreesurferSurfaceGeometryMetaInfo}}
}
\value{
a class of type \code{BrainSurface}
}
\description{
load Freesurfer ascii surface
}
\details{
requires rgl library
}

