% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\name{loadSurface}
\alias{loadSurface}
\title{load an surface from a set of files}
\usage{
loadSurface(surfaceName, surfaceDataName)
}
\arguments{
\item{surfaceName}{the name of the file containing the surface geometry.}

\item{surfaceDataName}{the name of the file containing the values to be mapped to the surface.}
}
\value{
an instance of the class \code{\linkS4class{BrainSurface}}
}
\description{
load an surface from a set of files
}

