\name{loadVolume}
\alias{loadVolume}
\title{load an image volume from a file}
\usage{
  loadVolume(fileName, index = 1)
}
\arguments{
  \item{fileName}{the name of the file to load}

  \item{index}{the index of the volume (e.g. if the file is
  4-dimensional)}
}
\description{
  load an image volume from a file
}

