\name{loadVolume}
\alias{loadVolume}

\title{ loads an image volume from file }
\description{
 loads an image volume from file and creates object of class "BrainVolume"
}
\usage{
loadVolume(file, volNum = 1)
}

\arguments{
  \item{file}{ the name of the image file to load }
  \item{volNum}{ the volume index if the file is 4 dimensional (starts at 1) }
}
\details{
  
}
\value{
  an object of class "BrainVolume"
}
\references{ }
\author{ }
\note{ 
}
\seealso{ }

\examples{}

\keyword{ manip }
