\name{loadVolumeList}
\alias{loadVolumeList}
\title{loadVolList...}
\usage{loadVolumeList(fileNames, mask)
}
\description{loadVolList}
\details{load a list of image volumes and return a \code{\linkS4class{BrainVector}} instance}
\value{an instance of class \code{\linkS4class{BrainVector}}}
\arguments{\item{fileNames}{a list of files to load}
}

