\name{lookup-methods}
\alias{lookup-methods}
\alias{lookup}
\title{Index Lookup operation...}
\description{Index Lookup operation}
\arguments{\item{x}{the object to query}
}
\alias{lookup,IndexLookupVolume,numeric-method}
\alias{lookup}
\alias{lookup,SparseBrainVector,numeric-method}
\alias{lookup}

