% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainVolume.R
\docType{methods}
\name{map}
\alias{map}
\alias{map,BrainVolume,Kernel-method}
\title{Generic function to apply a function to an object}
\usage{
map(x, m, ...)

\S4method{map}{BrainVolume,Kernel}(x, m, mask = NULL)
}
\arguments{
\item{x}{the object that is mapped}

\item{m}{the mapping object}

\item{...}{additional arguments}

\item{mask}{restrict application of kernel to masked area}
}
\description{
Generic function to apply a function to an object

apply a kernel function to a \code{\linkS4class{BrainVolume}}
}

