\name{map}
\alias{BrainVolume,}
\alias{BrainVolume-method}
\alias{Kernel,}
\alias{map}
\alias{map,}
\title{Generic function to apply a function to an object}
\arguments{
  \item{x}{the object that is mapped}

  \item{m}{the mapping object}
}
\description{
  Generic function to apply a function to an object

  apply a kernel function to a BrainVolume
}

