% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.R
\name{mapToColors}
\alias{mapToColors}
\title{mapToColors}
\usage{
mapToColors(imslice, col = heat.colors(128, alpha = 1),
  zero_col = "#00000000", alpha = 1, irange = range(imslice),
  threshold = c(0, 0))
}
\arguments{
\item{imslice}{vector or matrix of intensity values}

\item{col}{a color map}

\item{zero_col}{the background color.}

\item{alpha}{transparency multiplier}
}
\description{
map an matrix of intensity values to a matrix of color values.
}
