% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Axis.R
\name{matchAnatomy3D}
\alias{matchAnatomy3D}
\title{given three named axes return AxisSet3D singleton}
\usage{
matchAnatomy3D(axis1, axis2, axis3)
}
\arguments{
\item{axis1}{the first axis}

\item{axis2}{the second axis}

\item{axis3}{the third axis}
}
\description{
given three named axes return AxisSet3D singleton
}

