\name{matchAnatomy3D}
\alias{matchAnatomy3D}
\title{given three named axes return AxisSet3D singleton...}
\usage{matchAnatomy3D(axis1, axis2, axis3)
}
\description{given three named axes return AxisSet3D singleton}
\arguments{\item{axis1}{the first axis}
\item{axis2}{the second axis}
\item{axis3}{the thrid axis}
}

