% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainVolume.R
\docType{methods}
\name{mergePartitions}
\alias{mergePartitions}
\alias{mergePartitions,ClusteredBrainVolume,numeric,matrix-method}
\title{mergePartitions}
\usage{
mergePartitions(x, K, features, ...)

\S4method{mergePartitions}{ClusteredBrainVolume,numeric,matrix}(x, K, features)
}
\arguments{
\item{x}{the object to merge}

\item{K}{the number of merged partitions}

\item{features}{the features used to define the partition}

\item{...}{additional arguments}
}
\description{
mergePartitions

merge partititons in a ClusteredBrainVolume
}
