\name{mergePartitions}
\alias{mergePartitions}
\title{mergePartitions}
\arguments{
  \item{x}{the object to merge}

  \item{K}{the number of merged partitions}

  \item{features}{the features used to define the
  partition}
}
\description{
  mergePartitions

  merge partititons in a ClusteredBrainVolume
}

