% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{NamedAxis-class}
\alias{NamedAxis-class}
\title{This class represents an axis with a name attribute}
\description{
This class represents an axis with a name attribute
}
\section{Slots}{

\describe{
\item{\code{axis}}{the name of the axis}

\item{\code{direction}}{of axis (-1,+1)}
}}

