\name{NamedAxis-class}
\alias{NamedAxis-class}
\alias{NamedAxis}
\title{This class represents an axis with a name attribute...}
\description{This class represents an axis with a name attribute}
\section{Slots}{\describe{\item{\code{axis}:}{(\code{\link{character}}) the name of the axis}}}
\section{Methods}{\describe{\item{\code{\link[=show,NamedAxis-method]{show}}}{}\item{\code{\link[=print,NamedAxis-method]{print}}}{}}}

