\name{NamedAxis-class}
\alias{NamedAxis-class}
\alias{NamedAxis}
\title{This class supports reading of bulk binary data from a connection...}
\description{This class supports reading of bulk binary data from a connection}
\section{Slots}{\describe{\item{\code{axis}:}{(\code{\link{character}}) the name of the axis}}}
\section{Methods}{\describe{\item{\code{\link[=show,NamedAxis-method]{show}}}{}\item{\code{\link[=print,NamedAxis-method]{print}}}{}}}

