\docType{methods}
\name{ndim}
\alias{ndim}
\alias{ndim,AxisSet-method}
\alias{ndim,BrainData-method}
\alias{ndim,BrainSpace-method}
\title{Generic function to extract the number of dimensions of an object}
\usage{
ndim(x, ...)

\S4method{ndim}{AxisSet}(x)

\S4method{ndim}{BrainData}(x)

\S4method{ndim}{BrainSpace}(x)
}
\arguments{
  \item{x}{n-dimensional object}
}
\description{
Generic function to extract the number of dimensions of an
object

ndim

ndim
}

