% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Axis.R, R/BrainData.R, R/BrainSpace.R
\docType{methods}
\name{ndim}
\alias{ndim}
\alias{ndim,AxisSet-method}
\alias{ndim,BrainData-method}
\alias{ndim,BrainSpace-method}
\title{Generic function to extract the number of dimensions of an object}
\usage{
ndim(x, ...)

\S4method{ndim}{AxisSet}(x, ...)

\S4method{ndim}{BrainData}(x)

\S4method{ndim}{BrainSpace}(x)
}
\arguments{
\item{x}{n-dimensional object}

\item{...}{additional arguments}
}
\description{
Generic function to extract the number of dimensions of an object
}
\examples{
x = BrainSpace(c(10,10,10), c(1,1,1))
ndim(x) == 3
x = BrainSpace(c(10,10,10,3), c(1,1,1,1))
ndim(x) == 4
}

