\name{ndim-methods}
\alias{ndim-methods}
\alias{ndim}
\title{Generic function to extract the number of dimensions of an object...}
\description{Generic function to extract the number of dimensions of an object}
\arguments{\item{x}{n-dimensional object}
}
\alias{ndim,AxisSet-method}
\alias{ndim}
\alias{ndim,BrainData-method}
\alias{ndim}
\alias{ndim,BrainSpace-method}
\alias{ndim}

