% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrainMetaInfo.R
\name{NIfTIMetaInfo}
\alias{NIfTIMetaInfo}
\title{Constructor for \code{\linkS4class{NIfTIMetaInfo}} class}
\usage{
NIfTIMetaInfo(descriptor, nifti_header)
}
\arguments{
\item{descriptor}{an instance of class \code{\linkS4class{NIfTIFileDescriptor}}}

\item{nifti_header}{a \code{list} returned by \code{readNIftiHeader}}
}
\value{
an instance of class \code{\linkS4class{NIfTIMetaInfo}}
}
\description{
Constructor for \code{\linkS4class{NIfTIMetaInfo}} class
}

