\name{NIfTIMetaInfo-class}
\alias{NIfTIMetaInfo-class}
\alias{NIfTIMetaInfo}
\title{This class contains meta information for a NIfTI image file...}
\description{This class contains meta information for a NIfTI image file}
\section{Slots}{\describe{\item{\code{nifti_header}:}{(\code{\link{list}}) a list of attributes specific to the NIfTI file format}}}
\section{Extends}{\code{\linkS4class{BrainMetaInfo}}}

