\name{NIfTIMetaInfo-class}
\alias{NIfTIMetaInfo-class}
\alias{NIfTIMetaInfo}
\title{This class contains meta information for a NIfTI image file...}
\description{This class contains meta information for a NIfTI image file}
\section{Slots}{\describe{\item{\code{nifti_header}:}{}}}
\section{Extends}{\code{\linkS4class{BrainMetaInfo}}}
\section{Methods}{\describe{\item{\code{\link[=dataReader,NIfTIMetaInfo-method]{dataReader}}}{}}}

