% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainVolume.R
\docType{methods}
\name{numClusters}
\alias{numClusters}
\alias{numClusters,ClusteredBrainVolume-method}
\title{numClusters}
\usage{
numClusters(x)

\S4method{numClusters}{ClusteredBrainVolume}(x)
}
\arguments{
\item{x}{the object to extract number of clusters}
}
\description{
numClusters

get number of clusters in a ClusteredBrainVolume
}

