% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{overlay}
\alias{overlay}
\alias{overlay,Layer,Layer-method}
\title{overlay two objects}
\usage{
overlay(x, y, ...)

\S4method{overlay}{Layer,Layer}(x, y)
}
\arguments{
\item{x}{the underlay object}

\item{y}{the overlay object}

\item{...}{additional arguments for class-specific implementations}
}
\description{
overlay two objects

overlay
}

