% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{overlay}
\alias{overlay}
\title{overlay two objects}
\usage{
overlay(x, y, ...)
}
\arguments{
\item{x}{the underlay object}

\item{y}{the overlay object}

\item{...}{additional arguments for class-specific implementations}
}
\description{
overlay two objects
}
