% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Display.R
\docType{methods}
\name{overlay}
\alias{overlay}
\alias{overlay,Layer,Layer-method}
\alias{+,Overlay,Layer-method}
\alias{+,Layer,Layer-method}
\title{overlay two objects}
\usage{
overlay(x, y, ...)

\S4method{overlay}{Layer,Layer}(x, y)

\S4method{+}{Overlay,Layer}(e1, e2)

\S4method{+}{Layer,Layer}(e1, e2)
}
\arguments{
\item{x}{the underlay object}

\item{y}{the overlay object}

\item{...}{additional arguments for class-specific implementations}

\item{e1}{the left operand}

\item{e2}{the right operand}

\item{x}{the x layer}

\item{y}{the y layer}
}
\description{
overlay two objects

overlay
}
