\name{partition}
\alias{partition}
\alias{partition,ClusteredBrainVolume,ClusteredBrainVolume-method}
\title{partition}
\arguments{
  \item{x}{the object to partition}

  \item{K}{the number of partitions}

  \item{features}{the features used to define the
  partition}
}
\description{
  partition

  partition a ClusteredBrainVolume into K spatial disjoint
  components for every existing partition in the volume
}

