\docType{methods}
\name{permMat}
\alias{permMat}
\alias{permMat,AxisSet2D-method}
\alias{permMat,AxisSet3D-method}
\title{extract permutation matrix}
\usage{
permMat(x, ...)

\S4method{permMat}{AxisSet2D}(x)

\S4method{permMat}{AxisSet3D}(x)
}
\arguments{
  \item{x}{the object}
}
\description{
extract permutation matrix

permMat

permMat
}

