% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{permMat}
\alias{permMat}
\alias{permMat,AxisSet2D-method}
\title{extract permutation matrix}
\usage{
permMat(x, ...)

\S4method{permMat}{AxisSet2D}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional arguments}
}
\description{
extract permutation matrix

permMat
}

