% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Axis.R
\docType{methods}
\name{permMat}
\alias{permMat}
\alias{permMat,AxisSet2D-method}
\title{Extract permutation matrix}
\usage{
permMat(x, ...)

\S4method{permMat}{AxisSet2D}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additional arguments}
}
\description{
Extract permutation matrix

permMat
}

