% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{print}
\alias{print}
\title{Generic function to print an object}
\usage{
print(x, ...)
}
\arguments{
\item{x}{the object to print}

\item{...}{additional arguments}
}
\description{
Generic function to print an object
}

