% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Axis.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,AxisSet2D-method}
\alias{print,AxisSet3D-method}
\alias{print,NamedAxis-method}
\title{Generic function to print an object}
\usage{
print(x, ...)

\S4method{print}{NamedAxis}(x, ...)

\S4method{print}{AxisSet2D}(x, ...)

\S4method{print}{AxisSet3D}(x, ...)
}
\arguments{
\item{x}{the object to print}

\item{...}{additional arguments}
}
\description{
Generic function to print an object

print a \code{NamedAxis} instance

print a \code{AxisSet2D} instance

print a \code{AxisSet3D} instance
}

