% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainRegion3D.R
\name{RandomSearchlight}
\alias{RandomSearchlight}
\title{Create an spherical random searchlight iterator}
\usage{
RandomSearchlight(mask, radius)
}
\arguments{
\item{mask}{an volumetric image mask of type \code{\linkS4class{BrainVolume}} containing valid searchlight voxel set.}

\item{radius}{in mm of spherical searchlight}
}
\description{
Create an spherical random searchlight iterator
}

