% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BrainRegion3D.R
\name{RandomSearchlight}
\alias{RandomSearchlight}
\title{Create an Random Searchlight iterator}
\usage{
RandomSearchlight(mask, radius)
}
\arguments{
\item{mask}{an image volume containing valid central voxels for roving searchlight}

\item{radius}{in mm of spherical searchlight}
}
\description{
Create an Random Searchlight iterator
}

