\name{readElements}
\alias{readElements}
\title{Generic function to read a sequence of elements from an input source}
\arguments{
  \item{x}{the input channel}

  \item{numElements}{the number of elements to read}
}
\value{
  the elements as a vector
}
\description{
  Generic function to read a sequence of elements from an
  input source

  readElements
}

