\name{readElements-methods}
\alias{readElements-methods}
\alias{readElements}
\title{Generic function to read a sequence of elements from an input source...}
\description{Generic function to read a sequence of elements from an input source}
\value{\code{readElements-methods}: the elements as a vector

}
\arguments{\item{x}{the input channel}
\item{numElements}{the number of elements to read}
}
\alias{readElements,BinaryReader,numeric-method}
\alias{readElements}

