\name{readElements-methods}
\alias{readElements-methods}
\alias{readElements}
\title{Generic function to read a sequence of elements from an input source...}
\description{Generic function to read a sequence of elements from an input source}
\value{the elemtns as a vector}
\arguments{\item{x}{the input source}
\item{numElements}{the number of elements to be read}
}

