\name{readHeader}
\alias{readHeader}

\title{ reads the header part of an image file }
\description{
  a function to read an image header without loading associated image data
}
\usage{
readHeader(x)
}

\arguments{
  \item{x}{ an object derived from class "BrainFile" }
}

\value{
  an object derived from class "BrainHeader"
}

\examples{

}
\keyword{ manip }

