\name{readMetaInfo}
\alias{readMetaInfo}
\title{Generic function to read image meta info given a file and a \code{\linkS4class{BrainFileDescriptor}} instance.}
\arguments{
  \item{x}{descriptor instance}

  \item{fileName}{file name contianing meta information}
}
\description{
  Generic function to read image meta info given a file and
  a \code{\linkS4class{BrainFileDescriptor}} instance.
}

