\name{readMetaInfo-methods}
\alias{readMetaInfo-methods}
\alias{readMetaInfo}
\title{Generic function to read image meta information from a file...}
\description{Generic function to read image meta information from a file}
\value{\code{\linkS4class{BrainMetaInfo}} instance.}
\arguments{\item{x}{descriptor instance}
\item{fileName}{file name to be stripped of its extension}
}

