\name{readMetaInfo-methods}
\alias{readMetaInfo-methods}
\alias{readMetaInfo}
\title{Generic function to read image meta info given a file and a \linkS4class{BrainFileDescriptor} instance.}
\description{Generic function to read image meta info given a file and a \code{\linkS4class{BrainFileDescriptor}} instance.}
\arguments{\item{x}{descriptor instance}
\item{fileName}{file name contianing meta information}
}
\alias{readMetaInfo,NIfTIFileDescriptor-method}
\alias{readMetaInfo}
\alias{readMetaInfo,AFNIFileDescriptor-method}
\alias{readMetaInfo}

