\name{RegionSphere}
\alias{RegionSphere}
\title{Create A Spherical Region of interest}
\usage{
  RegionSphere(bvol, centroid, radius)
}
\arguments{
  \item{bvol}{an image volume}

  \item{centroid}{the center of the sphere in voxel space}

  \item{radius}{the radius of the spherical ROI}
}
\description{
  Create A Spherical Region of interest
}

