\name{RegionSphere}
\Rdversion{1.1}
\alias{RegionSphere}
\title{
	creation function for Spheric BrainRegion3D
}
\description{
	creates a new BrainRegion3D instance with a spheric geometry.
}

\usage{
	RegionSphere(bvol, centroid, radius)	
}

\arguments{
	\item{bvol}{the \code{BrainVolume} that on which the geometry of the region is based }
	\item{centroid}{the center coordinates (in grid space) of the region }
	\item{radius}{the radius of the sphere of the constructed region }

}


