% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ROIVolume-class}
\alias{ROIVolume-class}
\title{ROIVolume}
\description{
A class that represents a volumetric region of interest
}
\section{Slots}{

\describe{
\item{\code{data}}{the \code{numeric} or \code{matrix} data stored in ROI}

\item{\code{coords}}{the voxel coordinates of the ROI}
}}

