% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ROIVolume-class}
\alias{ROIVolume-class}
\title{ROIVolume}
\description{
A class that representing a volumetric region of interest (ROI).
}
\section{Slots}{

\describe{
\item{\code{data}}{the \code{numeric} data stored in the ROI}

\item{\code{coords}}{the voxel coordinates of the ROI}
}}

