% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainRegion3D.R
\name{ROIVolume}
\alias{ROIVolume}
\title{Create an instance of class \code{\linkS4class{ROIVolume}}}
\usage{
ROIVolume(vspace, coords, data = rep(nrow(coords), 1))
}
\arguments{
\item{coords}{matrix of voxel coordinates}

\item{data}{the data values, numeric vector}

\item{space}{an instance of class \code{BrainSpace}}
}
\value{
an instance of class \code{ROIVolume}
}
\description{
Create an instance of class \code{\linkS4class{ROIVolume}}
}

