\name{ROIVolume}
\alias{ROIVolume}
\title{Create an instance of class ROIVolume}
\usage{
  ROIVolume(vspace, coords, data = rep(length(indices), 1))
}
\arguments{
  \item{vspace}{the volume space}

  \item{coords}{matrix of coordinates}

  \item{data}{the data values}
}
\description{
  Create an instance of class ROIVolume
}

