% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Searchlight.R
\name{Searchlight}
\alias{Searchlight}
\title{Create an exhaustive searchlight iterator}
\usage{
Searchlight(mask, radius)
}
\arguments{
\item{mask}{an image volume containing valid central voxels for roving searchlight}

\item{radius}{in mm of spherical searchlight}
}
\value{
an \code{iter} class
}
\description{
Create an exhaustive searchlight iterator
}

