% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/SparseBrainVector.R,
%   R/BrainVector.R
\docType{methods}
\name{series_roi}
\alias{series_roi}
\alias{series}
\alias{series,SparseBrainVector,matrix-method}
\alias{series,SparseBrainVector,numeric-method}
\alias{series,BrainVector,matrix-method}
\alias{series_roi,BrainVector,matrix-method}
\alias{series,BrainVector,ROIVolume-method}
\alias{series_roi,BrainVector,ROIVolume-method}
\alias{series,BrainVector,LogicalBrainVolume-method}
\alias{series_roi,BrainVector,LogicalBrainVolume-method}
\alias{series,BrainVector,numeric-method}
\alias{series_roi,BrainVector,numeric-method}
\title{Extract one or more series from object and return as ROI object}
\usage{
series_roi(x, i, ...)

series(x, i, ...)

\S4method{series}{SparseBrainVector,matrix}(x, i)

\S4method{series}{SparseBrainVector,numeric}(x, i, j, k)

\S4method{series}{BrainVector,matrix}(x, i)

\S4method{series_roi}{BrainVector,matrix}(x, i)

\S4method{series}{BrainVector,ROIVolume}(x, i)

\S4method{series_roi}{BrainVector,ROIVolume}(x, i)

\S4method{series}{BrainVector,LogicalBrainVolume}(x, i)

\S4method{series_roi}{BrainVector,LogicalBrainVolume}(x, i)

\S4method{series}{BrainVector,numeric}(x, i, j, k)

\S4method{series_roi}{BrainVector,numeric}(x, i, j, k)
}
\arguments{
\item{x}{the object}

\item{i}{the series indices}

\item{...}{additional arguments}

\item{j}{index for 2nd dimension}

\item{k}{index for 3rd dimension}

\item{x}{the object}

\item{i}{the series indices}

\item{...}{additional arguments}
}
\description{
Extract one or more series from object and return as ROI object

Extract one or more series from object
}
