\name{series-methods}
\alias{series-methods}
\alias{series}
\title{Extract vector series from object...}
\description{Extract vector series from object}
\arguments{\item{x}{the object}
}
\alias{series,-method}
\alias{series}
\alias{series,-method}
\alias{series}
\alias{series,SparseBrainVector,matrix-method}
\alias{series}
\alias{series,-method}
\alias{series}

