\docType{methods}
\name{seriesIter}
\alias{seriesIter}
\alias{seriesIter,BrainVector-method}
\alias{seriesIter,SparseBrainVector-method}
\title{Construct a series iterator}
\usage{
seriesIter(x)

\S4method{seriesIter}{BrainVector}(x)

\S4method{seriesIter}{SparseBrainVector}(x)
}
\arguments{
  \item{x}{the object to be iterated}
}
\description{
Construct a series iterator
}

