\name{seriesIter-methods}
\alias{seriesIter-methods}
\alias{seriesIter}
\title{Construct a series iterator...}
\description{Construct a series iterator}
\arguments{\item{x}{the object to be iterated}
}
\alias{seriesIter,BrainVector-method}
\alias{seriesIter}
\alias{seriesIter,SparseBrainVector-method}
\alias{seriesIter}

