% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Axis.R
\docType{methods}
\name{show,AxisSet3D-method}
\alias{show,AxisSet3D-method}
\title{show an \code{AxisSet3D}}
\usage{
\S4method{show}{AxisSet3D}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show an \code{AxisSet3D}
}
