% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSpace.R
\docType{methods}
\name{show,BrainSpace-method}
\alias{show,BrainSpace-method}
\title{show a \code{BrainSpace}}
\usage{
\S4method{show}{BrainSpace}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show a \code{BrainSpace}
}
