% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainVector.R
\docType{methods}
\name{show,BrainVector-method}
\alias{show,BrainVector-method}
\title{show a \code{BrainVector}}
\usage{
\S4method{show}{BrainVector}(object)
}
\arguments{
\item{object}{the object}
}
\description{
show a \code{BrainVector}
}
