% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{show,BrainVector-method}
\alias{show,BrainVector-method}
\alias{show,BrainVolume-method}
\title{show}
\usage{
\S4method{show}{BrainVector}(object)

\S4method{show}{BrainVolume}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
show

show
}

