% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BrainVector.R, R/BrainVolume.R
\docType{methods}
\name{show,BrainVector-method}
\alias{show,BrainVector-method}
\alias{show,BrainVolume-method}
\title{show}
\usage{
\S4method{show}{BrainVector}(object)

\S4method{show}{BrainVolume}(object)
}
\description{
show

show
}

