\name{show-methods}
\alias{show,NamedAxis-method}
\alias{show}
\title{show,NamedAxis-method}
\alias{show,AxisSet1D-method}
\alias{show}
\alias{show,AxisSet2D-method}
\alias{show}
\alias{show,AxisSet3D-method}
\alias{show}
\alias{show,AxisSet4D-method}
\alias{show}
\alias{show,BrainVectorSource-method}
\alias{show}
\alias{show,BrainVector-method}
\alias{show}

