\name{slice}
\alias{slice}
\title{extract a 2D slice from an image volume}
\arguments{
  \item{x}{the object}

  \item{zlevel}{coordinate (in voxel units) along the
  sliced axis}

  \item{along}{the axis along which to slice}

  \item{orientation}{the target orientation of the 2D
  slice}
}
\description{
  extract a 2D slice from an image volume

  slice
}

