\docType{methods}
\name{space}
\alias{space}
\alias{space,BrainData-method}
\alias{space,IndexLookupVolume-method}
\title{Generic function to extract the \code{space} member variable}
\usage{
space(x, ...)

\S4method{space}{BrainData}(x)

\S4method{space}{IndexLookupVolume}(x)
}
\arguments{
  \item{x}{the object to query}

  \item{...}{additional arguments}
}
\value{
an object representing the geometric space of the image
}
\description{
Generic function to extract the \code{space} member
variable

space

space
}

