\name{space}
\alias{space}
\title{Generic function to extract the \code{space} member variable}
\arguments{
  \item{x}{the object to query}
}
\value{
  an object representing the geometric space of the image
}
\description{
  Generic function to extract the \code{space} member
  variable

  space

  space
}

