\name{space-methods}
\alias{space-methods}
\alias{space}
\title{Generic function to extract the space member variable...}
\description{Generic function to extract the \code{space} member variable}
\value{an object representing the geometric space of the image}
\arguments{\item{x}{the object to query}
}

