% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainData.R, R/BrainSpace.R
\docType{methods}
\name{spacing}
\alias{spacing}
\alias{spacing,BrainData-method}
\alias{spacing,BrainSpace-method}
\title{Generic function to extract the voxel dimensions of an image}
\usage{
spacing(x)

\S4method{spacing}{BrainData}(x)

\S4method{spacing}{BrainSpace}(x)
}
\arguments{
\item{x}{the object}
}
\value{
a numeric vector
}
\description{
Generic function to extract the voxel dimensions of an image

spacing
}
\examples{
bspace <- BrainSpace(c(10,10,10), c(2,2,2))
all.equal(spacing(bspace), c(2,2,2))
}

