\name{spacing-methods}
\alias{spacing-methods}
\alias{spacing}
\title{Generic function to extract the voxel dimensions of an image...}
\description{Generic function to extract the voxel dimensions of an image}
\value{\code{spacing-methods}: a numeric vector

}
\arguments{\item{x}{the object}
}
\alias{spacing,BrainData-method}
\alias{spacing}
\alias{spacing,BrainSpace-method}
\alias{spacing}

