% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClass.R, R/SparseBrainVector.R
\docType{class}
\name{SparseBrainVector-class}
\alias{SparseBrainVector}
\alias{SparseBrainVector-class}
\alias{indices,SparseBrainVector-method}
\title{SparseBrainVector}
\usage{
SparseBrainVector(data, space, mask, source = NULL, label = "")

\S4method{indices}{SparseBrainVector}(x)
}
\arguments{
\item{data}{an array which can be a \code{matrix} or 4-D \code{array}}

\item{space}{a BrainSpace instance}

\item{mask}{a 3D \code{array} of type \code{logical}}

\item{source}{the data source -- an instance of class \code{\linkS4class{BrainSource}}}

\item{label}{associated sub-image labels}
}
\description{
a sparse four-dimensional brain image, backed by a \code{matrix}, where each column represents
a vector spanning the fourth dimension (e.g. time)

constructs a SparseBrainVector object

return 1D index vector of location of nonzero elements in \code{SparseBrainVector}
}
\section{Slots}{

\describe{
\item{\code{mask}}{the mask defining the sparse domain}

\item{\code{data}}{the matrix of series, where rows span across voxel space and columns span the fourth dimensions}

\item{\code{map}}{instance of class \code{\linkS4class{IndexLookupVolume}} is used to map between spatial and index/row coordinates}
}}
\examples{
bspace <- BrainSpace(c(10,10,10,100), c(1,1,1))
mask <- array(rnorm(10*10*10) > .5, c(10,10,10))
mat <- matrix(rnorm(sum(mask)), 100, sum(mask))
svec <- SparseBrainVector(mat, bspace,mask)
length(indices(svec)) == sum(mask)
}

