\name{SparseBrainVector-class}
\alias{SparseBrainVector-class}
\alias{SparseBrainVector}
\title{a sparse Four-dimensional brain image, backed by a matrix, where each column represents...}
\description{a sparse Four-dimensional brain image, backed by a \code{matrix}, where each column represents 
a vector spanning the fourth dimension (e.g. time)}
\section{Extends}{\code{\linkS4class{BrainVector}}}
\section{Methods}{\describe{\item{\code{\link[=indices,SparseBrainVector-method]{indices}}}{}\item{\code{\link[=coords,SparseBrainVector-method]{coords}}}{}\item{\code{\link[=eachSeries,SparseBrainVector,function-method]{eachSeries}}}{}\item{\code{\link[=seriesIter,SparseBrainVector-method]{seriesIter}}}{}\item{\code{\link[=series,SparseBrainVector,matrix-method]{series}}}{}\item{\code{\link[=series,SparseBrainVector,numeric-method]{series}}}{}\item{\code{\link[=concat,SparseBrainVector,SparseBrainVector-method]{concat}}}{}\item{\code{\link[=concat,SparseBrainVector,SparseBrainVector-method]{concat}}}{}\item{\code{\link[=lookup,SparseBrainVector,numeric-method]{lookup}}}{}\item{\code{\link[=[,SparseBrainVector,numeric,numeric-method]{[}}}{}\item{\code{\link[=takeVolume,SparseBrainVector,numeric-method]{takeVolume}}}{}\item{\code{\link[=writeVector,SparseBrainVector,character-method]{writeVector}}}{}}}

