\name{SparseBrainVectorSource}
\alias{SparseBrainVectorSource}
\title{SparseBrainVectorSource}
\usage{
  SparseBrainVectorSource(metaInfo, indices, mask)
}
\arguments{
  \item{metaInfo}{an object of class
  \code{\linkS4class{BrainMetaInfo}}}

  \item{indices}{a vector of 1D indices}

  \item{mask}{a 3D \code{array} of type \code{logical}}
}
\description{
  constructs a SparseBrainVectorSource object
}

