% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{SparseBrainVolume-class}
\alias{SparseBrainVolume}
\alias{SparseBrainVolume-class}
\title{SparseBrainVolume}
\usage{
SparseBrainVolume(data, space, source = NULL, label = "", indices = NULL)
}
\arguments{
\item{data}{a numeric vector}

\item{space}{an instance of class \code{\linkS4class{BrainSpace}}}

\item{source}{an instance of class \code{\linkS4class{BrainSource}}}

\item{label}{a \code{character} string}

\item{indices}{a 1-d index vector}
}
\value{
\code{\linkS4class{DenseBrainVolume}} instance
}
\description{
Three-dimensional brain image, backed by a \code{sparseVector} for \code{Matrix} package

Construct a \code{\linkS4class{SparseBrainVolume}} instance
}

