% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClass.R, R/BrainVolume.R
\docType{class}
\name{SparseBrainVolume-class}
\alias{SparseBrainVolume}
\alias{SparseBrainVolume-class}
\alias{[,SparseBrainVolume,matrix,missing,ANY-method}
\alias{[,SparseBrainVolume,missing,missing,ANY-method}
\alias{[,SparseBrainVolume,missing,numeric,ANY-method}
\alias{[,SparseBrainVolume,numeric,missing,missing-method}
\alias{[,SparseBrainVolume,numeric,numeric,ANY-method}
\title{SparseBrainVolume}
\usage{
SparseBrainVolume(data, space, indices = NULL, source = NULL, label = "")

\S4method{[}{SparseBrainVolume,numeric,numeric,ANY}(x, i, j, k, ...,
  drop = TRUE)

\S4method{[}{SparseBrainVolume,numeric,missing,missing}(x, i, j, k, ..., drop)

\S4method{[}{SparseBrainVolume,matrix,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{SparseBrainVolume,missing,missing,ANY}(x, i, j, k, ...,
  drop = TRUE)

\S4method{[}{SparseBrainVolume,missing,numeric,ANY}(x, i, j, k, ...,
  drop = TRUE)
}
\arguments{
\item{data}{a numeric vector}

\item{space}{an instance of class \code{\linkS4class{BrainSpace}}}

\item{indices}{a index vector indicating the 1-d coordinates of the data values}

\item{source}{an instance of class \code{\linkS4class{BrainSource}}}

\item{label}{a \code{character} string}

\item{i}{index for dimension 1}

\item{j}{index for dimension 2}

\item{k}{index for dimension 3}

\item{...}{additional arguments}

\item{i}{index for dimension 1}

\item{j}{index for dimension 2}

\item{k}{index for dimension 3}

\item{...}{additional arguments}

\item{i}{matrix of voxel coordinates}

\item{i}{index for dimension 1}

\item{j}{index for dimension 2}

\item{k}{index for dimension 3}

\item{...}{additional arguments}

\item{i}{index for dimension 1}

\item{j}{index for dimension 2}

\item{k}{index for dimension 3}

\item{...}{additional arguments}
}
\value{
\code{\linkS4class{DenseBrainVolume}} instance
}
\description{
Three-dimensional brain image, backed by a \code{sparseVector} for \code{Matrix} package

Construct a \code{\linkS4class{SparseBrainVolume}} instance

extract data from \code{\linkS4class{SparseBrainVolume}}

extract data from \code{\linkS4class{SparseBrainVolume}}

extract data from \code{\linkS4class{SparseBrainVolume}}

extract data from \code{\linkS4class{SparseBrainVolume}}

extract data from \code{\linkS4class{SparseBrainVolume}}
}
\examples{
data <- 1:10
indices = seq(1,1000, length.out=10)
bspace <- BrainSpace(c(64,64,64), spacing=c(1,1,1))
sparsevol <- SparseBrainVolume(data,bspace,indices=indices)
densevol <- BrainVolume(data,bspace,indices=indices)
sum(sparsevol) == sum(densevol)
}

