% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{splitReduce}
\alias{splitReduce}
\alias{splitReduce,matrix,factor,function-method}
\title{Generic function to summarize subsets of an object}
\usage{
splitReduce(x, fac, FUN)

\S4method{splitReduce}{matrix,factor,`function`}(x, fac, FUN)
}
\arguments{
\item{x}{a numeric matrix(like) object}

\item{fac}{the factor to define subsets of the object}

\item{FUN}{the function to apply to each subset}
}
\value{
a new matrix(like) object where the original values have been scaled
}
\description{
Generic function to summarize subsets of an object
}

