\docType{methods}
\name{splitReduce}
\alias{splitReduce}
\title{Generic function to summarize subsets of an object}
\arguments{
  \item{x}{a numeric matrix(like) object}

  \item{fac}{the factor to define subsets of the object}

  \item{FUN}{the function to apply to each subset}
}
\value{
  a new matrix(like) object where the original values have
  been scaled
}
\description{
  Generic function to summarize subsets of an object

  splitScale
}

