\name{stripExtension-methods}
\alias{stripExtension-methods}
\alias{stripExtension}
\title{Generic function to strip extension from file name, given a \linkS4class{BrainFileDescriptor} instance.}
\description{Generic function to strip extension from file name, given a \code{\linkS4class{BrainFileDescriptor}} instance.}
\value{fileName without extension}
\arguments{\item{x}{descriptor instance}
\item{fileName}{file name to be stripped of its extension}
}

