\name{stripExtension}
\alias{stripExtension}
\title{Generic function to strip extension from file name, given a \code{\linkS4class{BrainFileDescriptor}} instance.}
\arguments{
  \item{x}{descriptor instance}

  \item{fileName}{file name to be stripped of its
  extension}
}
\value{
  fileName without extension
}
\description{
  Generic function to strip extension from file name, given
  a \code{\linkS4class{BrainFileDescriptor}} instance.
}

