% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseBrainVector.R
\docType{methods}
\name{[,SparseBrainVector,missing,missing,ANY-method}
\alias{[,SparseBrainVector,missing,missing,ANY-method}
\title{extractor}
\usage{
\S4method{[}{SparseBrainVector,missing,missing,ANY}(x, i, j, k, m, ...,
  drop = TRUE)
}
\arguments{
\item{x}{the object}

\item{i}{first index}

\item{j}{second index}

\item{k}{third index}

\item{m}{the fourth index}

\item{...}{additional args}

\item{drop}{dimension}
}
\description{
extractor
}
