% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SparseBrainVector.R
\docType{methods}
\name{[,SparseBrainVector,numeric,numeric,ANY-method}
\alias{[,SparseBrainVector,numeric,numeric,ANY-method}
\title{extract data from SparseBrainVector}
\usage{
\S4method{[}{SparseBrainVector,numeric,numeric,ANY}(x, i, j, k, m, ...,
  drop = TRUE)
}
\arguments{
\item{j}{index for dimension 2}

\item{k}{index for dimension 3}

\item{m}{index for dimension 4}

\item{...}{additional arguments}
}
\description{
extract data from SparseBrainVector
}

