% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainRegion3D.R
\name{SurfaceDisk}
\alias{SurfaceDisk}
\title{Create a Region on Surface}
\usage{
SurfaceDisk(surf, index, radius, max_order = NULL)
}
\description{
@param surf a \code{SurfaceGeometry} or \code{BrainSurface} or \code{BrainSurfaceVector}
 @param index the index of the central surface node
 @param radius the size in mm of the geodesic radius
 @param max_order maximum number of edges to traverse. 
   default is computed based on averaged edge length.
 @importFrom assertthat assert_that
 @rdname SurfaceDisk
}

