% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{takeSeries}
\alias{takeSeries}
\title{Generic function to extract a set of series from a 4D image}
\usage{
takeSeries(x, indices, ...)
}
\arguments{
\item{x}{a four dimensional image}

\item{indices}{the indices of the series' to extract}

\item{...}{additional arguments}
}
\description{
Generic function to extract a set of series from a 4D image
}
