\name{takeSeries-methods}
\alias{takeSeries-methods}
\alias{takeSeries}
\title{Generic function to extract a set of series from a 4D image...}
\description{Generic function to extract a set of series from a 4D image}
\arguments{\item{x}{a four dimensional image}
\item{i}{the indices of the series' to extract}
}

