\name{takeVolume-methods}
\alias{takeVolume-methods}
\alias{takeVolume}
\title{Generic function to extract a volume from a four-dimensional image...}
\description{Generic function to extract a volume from a four-dimensional image}
\arguments{\item{x}{four-dimensional image}
\item{i}{the indices of the volume(s) to extract}
}
\alias{takeVolume,BrainVector,numeric-method}
\alias{takeVolume}
\details{\code{takeVolume,BrainVector,numeric-method}: extract one or more volumes from a  BrainVector object

}
\alias{takeVolume,SparseBrainVector,numeric-method}
\alias{takeVolume}

