\name{takeVolume}
\alias{takeVolume}
\title{Generic function to extract a volume from a four-dimensional image}
\arguments{
  \item{x}{four-dimensional image}

  \item{i}{the indices of the volume(s) to extract}
}
\description{
  Generic function to extract a volume from a
  four-dimensional image

  takeVolume extract volume from SparseBrainVector

  takeVolume
}

