\name{takeVolume-methods}
\docType{methods}
\alias{takeVolume-methods}
\alias{takeVolume,BrainVector,numeric-method}
\alias{takeVolume,SparseBrainVector,numeric-method}
\alias{takeVolume}
\title{ Methods for Function takeVolume }
\description{
 	a function that extracts an image volume from 4-dimensional image
}
\section{Methods}{
\describe{

\item{x = "BrainVector", i = "numeric"}{ ~~describe this method here }

\item{x = "SparseBrainVector", i = "numeric"}{ ~~describe this method here }
}}
\keyword{methods}

