% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainVolume.R
\docType{methods}
\name{tesselate}
\alias{tesselate}
\alias{tesselate,LogicalBrainVolume,numeric-method}
\title{tesselate}
\usage{
tesselate(x, K, ...)

\S4method{tesselate}{LogicalBrainVolume,numeric}(x, K, features = NULL,
  spatialWeight = 4)
}
\arguments{
\item{x}{the object to tesselate}

\item{K}{the number of partitions}

\item{...}{extra arguments}

\item{features}{an optional matrix of additional features to tesselate volume}

\item{spatialWeight}{weight voxels according to distance}
}
\description{
tesselate

tesselate a LogicalBrainVolume into K spatial disjoint components using \code{kmeans} clustering.
}

