\name{tesselate}
\alias{tesselate}
\alias{tesselate,LogicalBrainVolume,LogicalBrainVolume-method}
\title{tesselate}
\arguments{
  \item{x}{the object to tesselate}

  \item{K}{the number of partitions}
}
\description{
  tesselate

  tesselate a LogicalBrainVolume into K spatial disjoint
  components
}

