\name{trans-methods}
\alias{trans-methods}
\alias{trans}
\title{Generic getter to extract image coordinate transformation...}
\description{Generic getter to extract image coordinate transformation}
\arguments{\item{x}{an object with a transformation}
}
\alias{trans,BrainMetaInfo-method}
\alias{trans}
\alias{trans,NIfTIMetaInfo-method}
\alias{trans}
\alias{trans,BrainSpace-method}
\alias{trans}
\alias{trans,BrainData-method}
\alias{trans}

