% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{values}
\alias{values}
\alias{values,ROIVolume-method}
\title{Generic function to extract data values of object}
\usage{
values(x, ...)

\S4method{values}{ROIVolume}(x)
}
\arguments{
\item{x}{the object to get values from}

\item{...}{ Arguments passed to \code{\link{sapply}} }
}
\description{
Generic function to extract data values of object

values
}

