% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainRegion3D.R
\docType{methods}
\name{values}
\alias{values}
\alias{values,ROIVolume-method}
\title{Generic function to extract data values of object}
\usage{
values(x, ...)

\S4method{values}{ROIVolume}(x)
}
\arguments{
\item{x}{the object to get values from}
}
\description{
Generic function to extract data values of object
}

