\name{voxels}
\alias{voxels}
\alias{voxels,Kernel,ANY-method}
\title{extract voxel coordinates}
\arguments{
  \item{x}{the object to extract voxels from}

  \item{kerndim}{the dimensions in voxels of the kernel}

  \item{vdim}{the dimensions of the voxels in real units}

  \item{FUN}{the kernel function taking as its first
  argument representing the distance from the center of the
  kernel}

  \item{...}{additional parameters to the kernel FUN}
}
\description{
  extract voxel coordinates

  extract voxels from a \code{Kernel} object
}

