\name{writeElements-methods}
\alias{writeElements-methods}
\alias{writeElements}
\title{Generic function to write a sequence of elements from an input source...}
\description{Generic function to write a sequence of elements from an input source}
\arguments{\item{x}{the output channel}
\item{els}{the elements to write}
}

