\name{writeElements}
\alias{writeElements}
\title{Generic function to write a sequence of elements from an input source}
\arguments{
  \item{x}{the output channel}

  \item{els}{the elements to write}
}
\description{
  Generic function to write a sequence of elements from an
  input source

  writeElements
}

