\name{writeVector-methods}
\alias{writeVector-methods}
\alias{writeVector}
\title{Generic function to write an image vector to disk...}
\description{Generic function to write an image vector to disk}
\arguments{\item{x}{the image to write}
}
\alias{writeVector,BrainVector,character,missing,missing-method}
\alias{writeVector}
\alias{writeVector,BrainVector,character,character,missing-method}
\alias{writeVector}

