\name{writeVector}
\alias{writeVector}
\alias{writeVector,BrainVector,character,character,missing,ANY-method}
\alias{writeVector,BrainVector,character,missing,character,ANY-method}
\alias{writeVector,BrainVector,character,missing,missing,ANY-method}
\title{Generic function to write an image vector to disk}
\arguments{
  \item{x}{the image to write}
}
\description{
  Generic function to write an image vector to disk

  writeVector

  writeVector

  writeVector
}

