\name{writeVector-methods}
\alias{writeVector-methods}
\alias{writeVector}
\title{Generic function to write an image vector to disk...}
\description{Generic function to write an image vector to disk}
\arguments{\item{x}{the image to write}
\item{fileName}{the output file name}
}
\alias{writeVector,BrainVector,character-method}
\alias{writeVector}
\alias{writeVector,SparseBrainVector,character-method}
\alias{writeVector}

