\name{writeVector-methods}
\docType{methods}
\alias{writeVector-methods}
\alias{writeVector,BrainVector,character,missing-method}
\alias{writeVector,SparseBrainVector,character,missing-method}
\alias{writeVector}
\title{ write a BrainVector instance to file }

\description{
 Methods for function \code{writeVector}  
}

\section{Methods}{
\describe{

\item{x = "BrainVector", fileName = "character", format = "missing"}{ write a BrainVector instance }

\item{x = "SparseBrainVector", fileName = "character", format = "missing"}{ write a SparseBrainVector instance }
}}

\keyword{methods}

