\name{writeVolume-methods}
\alias{writeVolume-methods}
\alias{writeVolume}
\title{Generic function to write an image volume to disk...}
\description{Generic function to write an image volume to disk}
\arguments{\item{x}{an image object}
\item{fileName}{a file name}
\item{format}{file format string}
\item{dataType}{output data type}
}
\alias{writeVector,BrainVector,character,missing,character-method}
\alias{writeVector}
\alias{writeVolume,BrainVolume,character,missing,missing-method}
\alias{writeVolume}
\alias{writeVolume,BrainVolume,character,character,missing-method}
\alias{writeVolume}
\alias{writeVolume,BrainVolume,character,missing,character-method}
\alias{writeVolume}

