\name{writeVolume}
\alias{writeVector}
\alias{writeVolume}
\title{Generic function to write an image volume to disk}
\arguments{
  \item{x}{an image object}

  \item{fileName}{a file name}

  \item{format}{file format string}

  \item{dataType}{output data type}
}
\description{
  Generic function to write an image volume to disk

  writeVector

  writeVolume

  writeVolume

  writeVolume
}

